Calculates paper statistics

Uses anonymous functions to calculate columns from `USA` sheet in WOODCARB spreadsheet.

calcUSApaper()

Value

A data frame with necessary intermediate calculations to calculate carbon from paper. Corresponds with necessary columns from the `USA` sheet in the WOODCARB spreadsheet (list them?)

Examples

calcUSApaper()

#>     Years apiTotalWP_L Recovered Fibre Pulp Exports Quantity
#> 1    1900     300848.1                                     0
#> 2    1901     334478.6                                     0
#> 3    1902     368109.2                                     0
#> 4    1903     401739.7                                     0
#> 5    1904     435370.3                                     0
#> 6    1905     482651.4                                     0
#> 7    1906     529932.6                                     0
#> 8    1907     577213.7                                     0
#> 9    1908     480040.5                                     0
#> 10   1909     565352.6                                     0
#> 11   1910     574064.0                                     0
#> 12   1911     608534.9                                     0
#> 13   1912     624167.8                                     0
#> 14   1913     639800.8                                     0
#> 15   1914     655433.7                                     0
#> 16   1915     652892.2                                     0
#> 17   1916     708796.4                                     0
#> 18   1917     724259.5                                     0
#> 19   1918     683799.8                                     0
#> 20   1919     654845.4                                     0
#> 21   1920     794260.3                                     0
#> 22   1921     597632.8                                     0
#> 23   1922     731899.2                                     0
#> 24   1923     787395.3                                     0
#> 25   1924     773802.1                                     0
#> 26   1925     823463.0                                     0
#> 27   1926     913359.1                                     0
#> 28   1927     896449.6                                     0
#> 29   1928     937474.3                                     0
#> 30   1929    1116099.0                                     0
#> 31   1930     879575.2                                     0
#> 32   1931     837600.8                                     0
#> 33   1932     714301.8                                     0
#> 34   1933     812309.4                                     0
#> 35   1934     842688.7                                     0
#> 36   1935     740855.4                                     0
#> 37   1936     810356.0                                     0
#> 38   1937     935241.2                                     0
#>     Paper+Paperboard Production Paper+Paperboard Exports
#> 1                       2286815                 72815.53
#> 2                       2469163                 72815.53
#> 3                       2651512                 72815.53
#> 4                       2833861                 72815.53
#> 5                       3016210                 72815.53
#> 6                       3213258                 75436.89
#> 7                       3410306                 74873.79
#> 8                       3607355                 74310.68
#> 9                       3804403                 73747.57
#> 10                      4001451                 73184.47
#> 11                      4201686                 72621.36
#> 12                      4401921                 78815.53
#> 13                      4602156                 85009.71
#> 14                      4802391                 91203.88
#> 15                      5002626                 97398.06
#> 16                      5160562                103592.23
#> 17                      5318498                164524.27
#> 18                      5476435                225456.31
#> 19                      5634371                286388.35
#> 20                      5792307                347320.39
#> 21                      6975847                408252.43
#> 22                      5178055                357087.38
#> 23                      6674596                162135.92
#> 24                      7641511                147475.73
#> 25                      7699015                141359.22
#> 26                      8739555                145145.63
#> 27                      9508821                147961.17
#> 28                      9710748                177766.99
#> 29                     10100328                178834.95
#> 30                     10815762                212330.10
#> 31                      9872951                254757.28
#> 32                      9108583                209514.56
#> 33                      7764924                165922.33
#> 34                      8922347                115825.24
#> 35                      8919027                123203.88
#> 36                     10173879                158446.60
#> 37                     11626750                173203.88
#> 38                     12463110                174466.02
#>     Pulp for Paper Production Pulp for Paper Imports Pulp for Paper Exports
#> 1                     1628822               100249.1               22693.57
#> 2                     1810901               130174.1               20112.91
#> 3                     1992980               160099.2               17532.26
#> 4                     2175059               190024.2               14951.60
#> 5                     2357138               219949.3               12370.95
#> 6                     2613123               209576.4               13324.79
#> 7                     2869108               244943.8               14278.64
#> 8                     3125093               364012.1               15232.48
#> 9                     2598988               307231.6               13856.30
#> 10                    3060876               453850.5               10981.27
#> 11                    3108040               621584.5               10255.16
#> 12                    3294669               689839.3               11644.83
#> 13                    3379307               662077.7               17403.47
#> 14                    3463945               664119.9               24256.19
#> 15                    3548584               828610.8               15131.91
#> 16                    3816968               685661.4               24481.57
#> 17                    4143797               824856.7               48281.56
#> 18                    4234199               817710.3               47264.61
#> 19                    3997661               697519.7               26930.45
#> 20                    4172797               754406.5               47513.37
#> 21                    4615964              1094651.7               38609.46
#> 22                    3473234               841977.5               34402.59
#> 23                    4253543              1520609.4               29593.01
#> 24                    4576067              1670615.7               27860.99
#> 25                    4497068              1839177.9               38771.31
#> 26                    4785680              2009361.0               45891.46
#> 27                    5308125              2091250.6               41331.90
#> 28                    5209853              2024040.9               38657.77
#> 29                    5448274              2119694.9               40442.94
#> 30                    5978984              2312380.9               66476.12
#> 31                    5509883              2177885.8               57625.02
#> 32                    5246945              1899677.8               63434.41
#> 33                    4474569              1763235.7               56951.50
#> 34                    5088513              2310398.6               94236.53
#> 35                    5278817              2149035.2              170083.42
#> 36                    5666524              2224023.3              197700.90
#> 37                    6505575              2601558.8              221015.41
#> 38                    7508159              2735326.3              368437.84
#>     Other Fibre Pulp Production Other Fibre Pulp Imports
#> 1                      300848.1                 18516.30
#> 2                      334478.6                 24043.54
#> 3                      368109.2                 29570.78
#> 4                      401739.7                 35098.03
#> 5                      435370.3                 40625.27
#> 6                      482651.4                 38709.36
#> 7                      529932.6                 45241.84
#> 8                      577213.7                 67234.09
#> 9                      480040.5                 56746.56
#> 10                     565352.6                 83827.51
#> 11                     574064.0                114808.45
#> 12                     608534.9                127415.33
#> 13                     624167.8                122287.67
#> 14                     639800.8                122664.87
#> 15                     655433.7                153046.82
#> 16                     652892.2                117282.36
#> 17                     708796.4                141091.72
#> 18                     724259.5                139869.33
#> 19                     683799.8                119310.73
#> 20                     654845.4                118390.52
#> 21                     794260.3                188354.66
#> 22                     597632.8                144877.49
#> 23                     731899.2                261648.41
#> 24                     787395.3                287459.71
#> 25                     773802.1                316463.89
#> 26                     823463.0                345746.97
#> 27                     913359.1                359837.56
#> 28                     896449.6                348272.92
#> 29                     937474.3                364731.93
#> 30                    1116099.0                431652.95
#> 31                     879575.2                347668.76
#> 32                     837600.8                303256.78
#> 33                     714301.8                281475.73
#> 34                     812309.4                368822.56
#> 35                     842688.7                343063.17
#> 36                     740855.4                290774.28
#> 37                     810356.0                324058.79
#> 38                     935241.2                340721.32
#>     Other Fibre Pulp Exports Wood Pulp for Paper Production
#> 1                   4191.568                        1327974
#> 2                   3714.913                        1476422
#> 3                   3238.259                        1624871
#> 4                   2761.605                        1773319
#> 5                   2284.950                        1921768
#> 6                   2461.128                        2130472
#> 7                   2637.306                        2339175
#> 8                   2813.484                        2547879
#> 9                   2559.299                        2118947
#> 10                  2028.273                        2495523
#> 11                  1894.157                        2533976
#> 12                  2150.835                        2686134
#> 13                  3214.472                        2755139
#> 14                  4480.188                        2824145
#> 15                  2794.907                        2893150
#> 16                  4187.572                        3164076
#> 17                  8258.559                        3435001
#> 18                  8084.610                        3509939
#> 19                  4606.453                        3313861
#> 20                  7456.367                        3517952
#> 21                  6643.457                        3821704
#> 22                  5919.589                        2875601
#> 23                  5092.014                        3521644
#> 24                  4793.988                        3788672
#> 25                  6671.306                        3723266
#> 26                  7896.457                        3962217
#> 27                  7111.903                        4394766
#> 28                  6651.770                        4313403
#> 29                  6958.941                        4510800
#> 30                 12409.120                        4862885
#> 31                  9199.023                        4630308
#> 32                 10126.409                        4409344
#> 33                  9091.505                        3760267
#> 34                 15043.534                        4276204
#> 35                 27151.420                        4436128
#> 36                 25847.903                        4925669
#> 37                 27530.413                        5695219
#> 38                 45893.839                        6572918
#>     Wood Pulp for Paper Imports Wood Pulp for Paper Exports
#> 1                       81732.8                    18502.00
#> 2                      106130.6                    16398.00
#> 3                      130528.4                    14294.00
#> 4                      154926.2                    12190.00
#> 5                      179324.0                    10086.00
#> 6                      170867.0                    10863.67
#> 7                      199702.0                    11641.33
#> 8                      296778.0                    12419.00
#> 9                      250485.0                    11297.00
#> 10                     370023.0                     8953.00
#> 11                     506776.0                     8361.00
#> 12                     562424.0                     9494.00
#> 13                     539790.0                    14189.00
#> 14                     541455.0                    19776.00
#> 15                     675564.0                    12337.00
#> 16                     568379.0                    20294.00
#> 17                     683765.0                    40023.00
#> 18                     677841.0                    39180.00
#> 19                     578209.0                    22324.00
#> 20                     636016.0                    40057.00
#> 21                     906297.0                    31966.00
#> 22                     697100.0                    28483.00
#> 23                    1258961.0                    24501.00
#> 24                    1383156.0                    23067.00
#> 25                    1522714.0                    32100.00
#> 26                    1663614.0                    37995.00
#> 27                    1731413.0                    34220.00
#> 28                    1675768.0                    32006.00
#> 29                    1754963.0                    33484.00
#> 30                    1880728.0                    54067.00
#> 31                    1830217.0                    48426.00
#> 32                    1596421.0                    53308.00
#> 33                    1481760.0                    47860.00
#> 34                    1941576.0                    79193.00
#> 35                    1805972.0                   142932.00
#> 36                    1933249.0                   171853.00
#> 37                    2277500.0                   193485.00
#> 38                    2394605.0                   322544.00
#>     Recovered Fibre Pulp Exports Recovered Paper Exports
#> 1                           0.00                13802.55
#> 2                           0.00                15345.48
#> 3                           0.00                16888.41
#> 4                           0.00                18431.34
#> 5                           0.00                19974.27
#> 6                           0.00                22143.48
#> 7                           0.00                24312.68
#> 8                           0.00                26481.88
#> 9                           0.00                22023.69
#> 10                          0.00                25937.71
#> 11                          0.00                26337.38
#> 12                          0.00                27918.86
#> 13                          0.00                28636.08
#> 14                          0.00                29353.30
#> 15                          0.00                30070.52
#> 16                          0.00                33976.50
#> 17                          0.00                36885.75
#> 18                          0.00                37690.45
#> 19                          0.00                35584.92
#> 20                          0.00                38988.46
#> 21                          0.00                49219.73
#> 22                          0.00                37034.87
#> 23                          0.00                45355.25
#> 24                          0.00                48794.31
#> 25                          0.00                47951.94
#> 26                          0.00                51029.39
#> 27                          0.00                56600.19
#> 28                          0.00                55552.32
#> 29                          0.00                58094.60
#> 30                          0.00                71364.22
#> 31                          0.00                64947.24
#> 32                          0.00                61847.88
#> 33                          0.00                52743.57
#> 34                          0.00                59980.39
#> 35                          0.00                62223.57
#> 36                          0.00                65894.68
#> 37                          0.00                72624.40
#> 38                          0.00                83816.66
#>     Total Roundwood Consumed For Paper Imported Pulpwood Chips
#> 1                             1406.322                0.000000
#> 2                             1562.580                0.000000
#> 3                             1666.752                0.000000
#> 4                             1823.010                0.000000
#> 5                             1979.268                0.000000
#> 6                             2031.354                0.000000
#> 7                             2343.870                0.000000
#> 8                             2448.042                0.000000
#> 9                             2135.526                0.000000
#> 10                            2395.956                0.000000
#> 11                            2291.784                0.000000
#> 12                            2500.128                0.000000
#> 13                            2604.300                0.000000
#> 14                            2708.472                0.000000
#> 15                            2760.558                0.000000
#> 16                            3125.160                0.000000
#> 17                            3385.591                0.000000
#> 18                            3593.935                0.000000
#> 19                            3489.763                0.000000
#> 20                            3437.677                0.000000
#> 21                            3750.193                0.000000
#> 22                            2708.472                0.000000
#> 23                            3541.849                0.000000
#> 24                            3541.849                0.000000
#> 25                            3541.849                0.000000
#> 26                            3593.935                0.000000
#> 27                            4166.881                0.000000
#> 28                            3958.537                0.000000
#> 29                            4166.881                0.000000
#> 30                            4635.655                0.000000
#> 31                            4114.795                0.000000
#> 32                            4166.881                0.000000
#> 33                            3646.021                0.000000
#> 34                            4323.139                0.000000
#> 35                            4479.397                0.000000
#> 36                            5052.343                0.000000
#> 37                            5781.547                0.000000
#> 38                            6667.009                0.000000
#>     Chip Exports (tons) Paper+Paperbaord Imports Paper+Paperboard Consumption
#> 1                   0.0                 9708.738                      2223708
#> 2                   0.0                14563.107                      2410911
#> 3                   0.0                19417.476                      2598114
#> 4                   0.0                24271.845                      2785317
#> 5                   0.0                29126.214                      2972520
#> 6                   0.0                33980.583                      3171802
#> 7                   0.0                38834.951                      3374268
#> 8                   0.0                43689.320                      3576733
#> 9                   0.0                48543.689                      3779199
#> 10                  0.0                54368.932                      3982636
#> 11                  0.0               111320.388                      4240385
#> 12                  0.0               168271.845                      4491378
#> 13                  0.0               225223.301                      4742370
#> 14                  0.0               282174.757                      4993362
#> 15                  0.0               339126.214                      5244354
#> 16                  0.0               408679.612                      5465650
#> 17                  0.0               478233.010                      5632207
#> 18                  0.0               547786.408                      5798765
#> 19                  0.0               617339.806                      5965322
#> 20                  0.0               686893.204                      6131880
#> 21                  0.0               798349.515                      7365944
#> 22                  0.0               835631.068                      5656599
#> 23                  0.0              1108737.864                      7621198
#> 24                  0.0              1425728.155                      8919763
#> 25                  0.0              1456699.029                      9014354
#> 26                  0.0              1521456.311                     10115866
#> 27                  0.0              1916019.417                     11276880
#> 28                  0.0              2045922.330                     11578903
#> 29                  0.0              2200776.699                     12122270
#> 30                  0.0              2459805.825                     13063238
#> 31                  0.0              2296407.767                     11914602
#> 32                  0.0              2080970.874                     10980039
#> 33                  0.0              1801262.136                      9400264
#> 34                  0.0              1798446.602                     10604968
#> 35                  0.0              2199320.388                     10995144
#> 36                  0.0              2385436.893                     12400869
#> 37                  0.0              2771941.748                     14225487
#> 38                  0.0              3313009.709                     15601653
#>     Percent of Wood Pulp For Paper CarbonInputFlowPaperStockChange
#> 1                        0.8152971                       0.7072256
#> 2                        0.8152971                       0.7667635
#> 3                        0.8152971                       0.8263013
#> 4                        0.8152971                       0.8858392
#> 5                        0.8152971                       0.9453770
#> 6                        0.8152971                       1.0087563
#> 7                        0.8152971                       1.0731483
#> 8                        0.8152971                       1.1375403
#> 9                        0.8152971                       1.2019323
#> 10                       0.8152971                       1.2666331
#> 11                       0.8152971                       1.3486074
#> 12                       0.8152971                       1.4284327
#> 13                       0.8152971                       1.5082580
#> 14                       0.8152971                       1.5880833
#> 15                       0.8152971                       1.6679086
#> 16                       0.8289500                       1.7673984
#> 17                       0.8289500                       1.8212572
#> 18                       0.8289500                       1.8751160
#> 19                       0.8289500                       1.9289749
#> 20                       0.8430680                       2.0166036
#> 21                       0.8279319                       2.3789610
#> 22                       0.8279319                       1.8268981
#> 23                       0.8279319                       2.4613998
#> 24                       0.8279319                       2.8807942
#> 25                       0.8279319                       2.9113442
#> 26                       0.8279319                       3.2670967
#> 27                       0.8279319                       3.6420664
#> 28                       0.8279319                       3.7396102
#> 29                       0.8279319                       3.9151001
#> 30                       0.8133297                       4.1445918
#> 31                       0.8403641                       3.9058121
#> 32                       0.8403641                       3.5994462
#> 33                       0.8403641                       3.0815687
#> 34                       0.8403641                       3.4764915
#> 35                       0.8403641                       3.6043978
#> 36                       0.8692575                       4.2049893
#> 37                       0.8754367                       4.8579855
#> 38                       0.8754367                       5.3279444
#>     Nonwood Fiber Percent of Total Pulp Consumption
#> 1                                       0.184702902
#> 2                                       0.184702902
#> 3                                       0.184702902
#> 4                                       0.184702902
#> 5                                       0.184702902
#> 6                                       0.184702902
#> 7                                       0.184702902
#> 8                                       0.184702902
#> 9                                       0.184702902
#> 10                                      0.184702902
#> 11                                      0.184702902
#> 12                                      0.184702902
#> 13                                      0.184702902
#> 14                                      0.184702902
#> 15                                      0.184702902
#> 16                                      0.171049972
#> 17                                      0.171049972
#> 18                                      0.171049972
#> 19                                      0.171049972
#> 20                                      0.156931990
#> 21                                      0.172068123
#> 22                                      0.172068123
#> 23                                      0.172068123
#> 24                                      0.172068123
#> 25                                      0.172068123
#> 26                                      0.172068123
#> 27                                      0.172068123
#> 28                                      0.172068123
#> 29                                      0.172068123
#> 30                                      0.186670345
#> 31                                      0.159635903
#> 32                                      0.159635903
#> 33                                      0.159635903
#> 34                                      0.159635903
#> 35                                      0.159635903
#> 36                                      0.130742463
#> 37                                      0.124563317
#> 38                                      0.124563317
#>     Imported Woodpulp as a Percent of Total WoodPulp Consumption
#> 1                                                     0.05874967
#> 2                                                     0.06776508
#> 3                                                     0.07496870
#> 4                                                     0.08085684
#> 5                                                     0.08575968
#> 6                                                     0.07459894
#> 7                                                     0.07901993
#> 8                                                     0.10478569
#> 9                                                     0.10622165
#> 10                                                    0.12953298
#> 11                                                    0.16712093
#> 12                                                    0.17363782
#> 13                                                    0.16453298
#> 14                                                    0.16183011
#> 15                                                    0.18995849
#> 16                                                    0.15311272
#> 17                                                    0.16764111
#> 18                                                    0.16339030
#> 19                                                    0.14941782
#> 20                                                    0.15460130
#> 21                                                    0.19299196
#> 22                                                    0.19668655
#> 23                                                    0.26470426
#> 24                                                    0.26863861
#> 25                                                    0.29205007
#> 26                                                    0.29772062
#> 27                                                    0.28421285
#> 28                                                    0.28130294
#> 29                                                    0.28159250
#> 30                                                    0.28114434
#> 31                                                    0.28543181
#> 32                                                    0.26819530
#> 33                                                    0.28527385
#> 34                                                    0.31629038
#> 35                                                    0.29610137
#> 36                                                    0.28910277
#> 37                                                    0.29276661
#> 38                                                    0.27699373
#>     Production Approach-C Input from Paper Products(Calc BU)
#> 1                                                  0.6988044
#> 2                                                  0.7459067
#> 3                                                  0.7934912
#> 4                                                  0.8414258
#> 5                                                  0.8896226
#> 6                                                  0.9595411
#> 7                                                  1.0139583
#> 8                                                  1.0433330
#> 9                                                  1.0954218
#> 10                                                 1.1221730
#> 11                                                 1.1272504
#> 12                                                 1.1723257
#> 13                                                 1.2408031
#> 14                                                 1.3010889
#> 15                                                 1.3064330
#> 16                                                 1.4360419
#> 17                                                 1.4647261
#> 18                                                 1.5146804
#> 19                                                 1.5741106
#> 20                                                 1.6441668
#> 21                                                 1.8524280
#> 22                                                 1.3712834
#> 23                                                 1.6142975
#> 24                                                 1.8348732
#> 25                                                 1.7941700
#> 26                                                 2.0200567
#> 27                                                 2.2364140
#> 28                                                 2.2907688
#> 29                                                 2.3819401
#> 30                                                 2.5205463
#> 31                                                 2.3605311
#> 32                                                 2.2340039
#> 33                                                 1.8621093
#> 34                                                 2.0599401
#> 35                                                 2.1486871
#> 36                                                 2.5553097
#> 37                                                 2.9226300
#> 38                                                 3.2536326
#>  [ reached getOption("max.print") -- omitted 83 rows ]