Calculate carbon in SWDS for the Production Method

Column names with only letters come from Calc sheet, otherwise sheet is specified (need to change names)

calculatedumpcarbonproduction(years, totalcarbonstockchange = FALSE,
  swpSwdsNondegradable = 0.77, paperSwdsNondegradable = 0.44,
  paperLandfillDecay = 0.0478032538317204,
  swpLandfillDecay = 0.0300063714528115, swpDumpDecay = 0.0420089200339361,
  paperDumpDecay = 0.0840178400678722, woodToCarbon = 4.535925e-07,
  paperToCarbon = 3.9008955e-07)

Arguments

years

years to calculate

totalcarbonstockchange

If true, returns carbon stock change, or returns data frame of intermediate calculations

swpSwdsNondegradable

Percent of solid-wood products in landfills and dumps that are not subject to decay.

paperSwdsNondegradable

Percent of paper products in landfills and dumps that are not subject to decay.

paperLandfillDecay

Decay rate of paper products in landfills as a half-life in years.

swpLandfillDecay

Decay rate of solid-wood products in landfills as a half-life in years.

swpDumpDecay

Decay rate of solid-wood products in dumps as a half-life in years.

paperDumpDecay

Decay rate of paper products in dumps as a half-life in years.

woodToCarbon

Conversion factor to convert oven-dry tons of solid wood to Tg of Carbon.

paperToCarbon

Conversion factor to convert oven-dry tons of paper to Tg of Carbon.

Value

Either a data frame of values from Calculation spreadsheet or total carbon output