Calculate carbon in SWDS for the Stock Change Method

Calculate carbon in SWDS for the Stock Change Method

calculatedumpcarbonstockchange(Years, totalcarbonstockchange = FALSE)

Arguments

Years

years to calculate

totalcarbonstockchange

Whether to return data frame of intermedate calculations or just totals

Value

If totalcarbonstockchange is false: a data frame of intermediate calculations. If totalcarbonstockchange is true: a numeric vector of stock change for total carbon from dumps.

Examples

calculatedumpcarbonstockchange(Years = 1990:2015)

#>     Years TotalCarbonOutputStockChange ParamResults_Q   Dumps_C   Dumps_D
#> 91   1990                     50789440     0.01517695 0.3496421 129.91531
#> 92   1991                     50653922     0.01638197 0.3763961 125.03896
#> 93   1992                     51521203     0.01490181 0.3482497 120.33218
#> 94   1993                     52205309     0.01503865 0.3561142 115.82271
#> 95   1994                     52891673     0.01517198 0.3639950 111.50260
#> 96   1995                     53352870     0.01613588 0.3904956 107.38209
#> 97   1996                     54077914     0.01596480 0.3916059 103.42877
#> 98   1997                     54787478     0.01575570 0.3915479  99.63476
#> 99   1998                     55587859     0.01570413 0.3959676  95.99796
#> 100  1999                     56476871     0.01599190 0.4096724  92.52093
#> 101  2000                     57095559     0.01611950 0.4174646  89.19155
#> 102  2001                     57475588     0.01612050 0.4202695  85.99909
#> 103  2002                     58306899     0.01622322 0.4290647  82.94378
#> 104  2003                     58878411     0.01606186 0.4289610  80.01154
#> 105  2004                     60241448     0.01603494 0.4381558  77.20634
#> 106  2005                     61080815     0.01605844 0.4449119  74.52072
#> 107  2006                     61324267     0.01605844 0.4466852  71.94507
#> 108  2007                     60853367     0.01605844 0.4432552  69.46997
#> 109  2008                     59250726     0.01605844 0.4315815  67.08345
#> 110  2009                     58232243     0.01605844 0.4241629  64.78602
#> 111  2010                     58163727     0.01605844 0.4236639  62.58073
#> 112  2011                     58298544     0.01605844 0.4246459  60.46530
#> 113  2012                     58507421     0.01605844 0.4261673  58.43661
#> 114  2013                     58959537     0.01605844 0.4294605  56.49286
#> 115  2014                     59492926     0.01605844 0.4333457  54.63121
#> 116  2015                     60054468     0.01605844 0.4374360  52.84854
#>            Q   Dumps_F        X        Y       AA        V        D         J
#> 91  17.13246 -5.107960 3.940466 49.98057 2.440731 13.19200 32.81423  97.07387
#> 92  18.12870 -4.876355 4.169601 52.57266 2.592087 13.95910 32.56279  98.57906
#> 93  16.49147 -4.706775 3.793037 54.72364 2.150980 12.69843 33.67451 100.56903
#> 94  16.58525 -4.509473 3.814608 56.83290 2.109259 12.77064 34.63855 102.81533
#> 95  16.67655 -4.320109 3.835608 58.90109 2.068199 12.84095 36.54295 105.97164
#> 96  17.60415 -4.120510 4.048955 61.11618 2.215081 13.55520 36.58388 108.40289
#> 97  17.37573 -3.953325 3.996417 63.21572 2.099543 13.37931 36.44060 110.14273
#> 98  17.10318 -3.794001 3.933730 65.19324 1.977518 13.16944 38.23753 112.83217
#> 99  17.03138 -3.636803 3.917218 67.09711 1.903878 13.11417 38.02705 114.71725
#> 100 17.35492 -3.477032 3.991631 69.01777 1.920659 13.36329 39.98446 117.63917
#> 101 17.42191 -3.329376 4.007039 70.89744 1.879664 13.41487 39.36994 119.39377
#> 102 17.53896 -3.192460 4.033962 72.74848 1.851044 13.50500 37.11494 119.01326
#> 103 17.90601 -3.055314 4.118382 74.62756 1.879080 13.78763 37.55138 119.05579
#> 104 17.90168 -2.932237 4.117387 76.45093 1.823372 13.78430 36.48277 118.27554
#> 105 18.28541 -2.805199 4.205643 78.30687 1.855938 14.07976 38.11726 118.92512
#> 106 18.56736 -2.685623 4.270492 80.17170 1.864830 14.29686 37.27240 118.77662
#> 107 18.64136 -2.575650 4.287513 81.99873 1.827029 14.35385 37.44107 118.79197
#> 108 18.49822 -2.475103 4.254590 83.74057 1.741839 14.24363 36.58191 118.15031
#> 109 18.01104 -2.386522 4.142540 85.32288 1.582310 13.86850 33.83168 115.57103
#> 110 17.70145 -2.297428 4.071333 86.78996 1.467081 13.63011 28.98965 109.92767
#> 111 17.68062 -2.205285 4.066542 88.20965 1.419691 13.61408 31.27297 107.37261
#> 112 17.72160 -2.115436 4.075968 89.59713 1.387483 13.64563 29.99253 104.45600
#> 113 17.78510 -2.028691 4.090572 90.95837 1.361241 13.69452 29.12277 101.57674
#> 114 17.92253 -1.943744 4.122182 92.31065 1.352274 13.80035 29.39888  99.59724
#> 115 18.08467 -1.861652 4.159474 93.65973 1.349085 13.92520 37.84014 104.51093
#> 116 18.25537 -1.782674 4.198734 95.00763 1.347900 14.05663 40.81893 110.51257
#>            K   Dumps_H  Dumps_I         R       AG    Dumps_K       AI
#> 91  30.58338 0.4641624 56.70510 16.481699 7.251948 -4.3000777 9.229751
#> 92  31.05760 0.5087847 52.77947 15.490235 6.815703 -3.9256323 8.674532
#> 93  31.68454 0.4721570 49.12431 15.339178 6.749238 -3.6551613 8.589940
#> 94  32.39225 0.4871355 45.76626 15.584132 6.857018 -3.3580469 8.727114
#> 95  33.38665 0.5065416 42.68638 14.844997 6.531799 -3.0798762 8.313198
#> 96  34.15262 0.5510825 39.88631 13.424399 5.906736 -2.8000788 7.517663
#> 97  34.70076 0.5539907 37.30593 12.748659 5.609410 -2.5803732 7.139249
#> 98  35.54808 0.5600849 34.93118 13.588551 5.978963 -2.3747570 7.609589
#> 99  36.14198 0.5675783 32.74739 14.091656 6.200329 -2.1837866 7.891328
#> 100 37.06254 0.5927005 30.75603 14.464991 6.364596 -1.9913550 8.100395
#> 101 37.61533 0.6063402 28.93160 14.079428 6.194948 -1.8244306 7.884480
#> 102 37.49545 0.6044455 27.24683 12.924957 5.686981 -1.6847742 7.237976
#> 103 37.50885 0.6085141 25.69639 13.105610 5.766468 -1.5504409 7.339141
#> 104 37.26303 0.5985136 24.25689 12.041804 5.298394 -1.4394979 6.743410
#> 105 37.46768 0.6007919 22.93106 11.826600 5.203704 -1.3258265 6.622896
#> 106 37.42090 0.6009212 21.70812 11.523835 5.070487 -1.2229479 6.453348
#> 107 37.42573 0.6009988 20.58003 12.144406 5.343538 -1.1280905 6.800867
#> 108 37.22357 0.5977525 19.53637 11.924872 5.246944 -1.0436515 6.677929
#> 109 36.41096 0.5847032 18.56157 11.513972 5.066148 -0.9748001 6.447824
#> 110 34.63301 0.5561520 17.63599 10.808521 4.755749 -0.9255856 6.052772
#> 111 33.82803 0.5432253 16.77022 10.417405 4.583658 -0.8657720 5.833747
#> 112 32.90914 0.5284694 15.95793 10.134432 4.459150 -0.8122817 5.675282
#> 113 32.00203 0.5139026 15.19517  9.855084 4.336237 -0.7627629 5.518847
#> 114 31.37838 0.5038878 14.48229  9.663031 4.251734 -0.7128828 5.411297
#> 115 32.92645 0.5287474 13.84759 10.139762 4.461495 -0.6346973 5.678267
#> 116 34.81728 0.5591112 13.29010 10.722048 4.717701 -0.5574941 6.004347
#>            AJ          AL Total_swpstock_change_LD Total_paper_stock_change_LD
#> 91   91.24896  4.86775434                 10.52477                    7.819624
#> 92   95.36475  4.11578640                 11.67483                    7.005857
#> 93   99.21203  3.84728184                 10.14263                    6.941359
#> 94  103.01470  3.80267574                 10.37043                    7.301647
#> 95  106.24886  3.23415722                 10.58904                    6.686080
#> 96  108.57623  2.32736649                 11.64977                    5.434023
#> 97  110.43626  1.86003616                 11.52553                    4.889073
#> 98  112.66032  2.22405888                 11.35296                    5.828265
#> 99  115.05180  2.39147717                 11.38124                    6.408019
#> 100 117.53370  2.48190146                 11.80691                    6.855142
#> 101 119.69631  2.16260692                 11.96516                    6.533125
#> 102 121.14324  1.44693480                 12.16359                    5.449142
#> 103 122.62071  1.47747226                 12.61139                    5.693500
#> 104 123.46223  0.84151394                 12.67543                    4.700410
#> 105 124.15033  0.68810592                 13.13050                    4.565983
#> 106 124.64523  0.49489982                 13.47607                    4.342439
#> 107 125.44922  0.80398625                 13.60523                    5.019434
#> 108 126.09920  0.64997667                 13.51036                    4.853269
#> 109 126.49991  0.40071698                 13.06429                    4.492065
#> 110 126.50532  0.00540598                 12.79977                    3.835570
#> 111 126.30145 -0.20387317                 12.82848                    3.514013
#> 112 125.95564 -0.34580733                 12.91768                    3.301061
#> 113 125.47631 -0.47932899                 13.02707                    3.094145
#> 114 124.91621 -0.56010381                 13.20888                    2.978747
#> 115 124.63645 -0.27976095                 13.41263                    3.547037
#> 116 124.68065  0.04420592                 13.62186                    4.204413
#>     Var1b_STOCKCHANGE_TOTAL
#> 91                 18.34439
#> 92                 18.68069
#> 93                 17.08399
#> 94                 17.67208
#> 95                 17.27512
#> 96                 17.08379
#> 97                 16.41460
#> 98                 17.18123
#> 99                 17.78926
#> 100                18.66206
#> 101                18.49828
#> 102                17.61273
#> 103                18.30489
#> 104                17.37584
#> 105                17.69648
#> 106                17.81851
#> 107                18.62466
#> 108                18.36363
#> 109                17.55636
#> 110                16.63534
#> 111                16.34250
#> 112                16.21874
#> 113                16.12122
#> 114                16.18763
#> 115                16.95966
#> 116                17.82627

calculatedumpcarbonstockchange(Years = 1990:2015, totalcarbonstockchange = TRUE)

#>  [1] 18.34439 18.68069 17.08399 17.67208 17.27512 17.08379 16.41460 17.18123
#>  [9] 17.78926 18.66206 18.49828 17.61273 18.30489 17.37584 17.69648 17.81851
#> [17] 18.62466 18.36363 17.55636 16.63534 16.34250 16.21874 16.12122 16.18763
#> [25] 16.95966 17.82627