<<<<<<< HEAD

Solidwood products statistics

>>>>>>> master
Various production, imports and export data for different wood product types. 21 columns in SWPcalcdata that are used to calculate production and stock change approach.

calculateswpdata()
<<<<<<< HEAD
Value

Data frame of Roundwood and solidwood production, imports and exports corresponds to columns in `SWPcalcdata` spreadsheet.

Examples
=======

Value

Data frame of Roundwood and solidwood production, imports and exports corresponds to columns in `SWPcalcdata` spreadsheet.

Examples
>>>>>>> master

calculateswpdata()

#>     Years Other Products Production Other Products Exports Sawnwood Production
#> 1    1900                  21749255                      0            34987067
#> 2    1901                  22196158                      0            36479140
#> 3    1902                  22717544                      0            37971213
#> 4    1903                  23462381                      0            39463286
#> 5    1904                  23834800                      0            40955359
#> 6    1905                  24430670                      0            41409956
#> 7    1906                  26814150                      0            43767172
#> 8    1907                  31432142                      0            43744381
#> 9    1908                  28750728                      0            39921801
#> 10   1909                  30538338                      0            42286345
#> 11   1910                  30910756                      0            42207669
#> 12   1911                  30091435                      0            40716540
#> 13   1912                  30016951                      0            42541370
#> 14   1913                  29719016                      0            41526144
#> 15   1914                  28676244                      0            38160639
#> 16   1915                  28080374                      0            34519723
#> 17   1916                  28750728                      0            37314594
#> 18   1917                  28750728                      0            33106186
#> 19   1918                  28601760                      0            29612632
#> 20   1919                  28527276                      0            32283978
#> 21   1920                  28154858                      0            32778060
#> 22   1921                  25622410                      0            26877959
#> 23   1922                  25249991                      0            32459841
#> 24   1923                  25398959                      0            37962696
#> 25   1924                  24654121                      0            36799572
#> 26   1925                  22270641                      0            37930985
#> 27   1926                  22196158                      0            36869309
#> 28   1927                  21376836                      0            34592856
#> 29   1928                  20631999                      0            33990112
#> 30   1929                  20557515                      0            36150677
#> 31   1930                  17801616                      0            27460182
#> 32   1931                  14449848                      0            18672402
#> 33   1932                  12364302                      0            12599299
#> 34   1933                  12438786                      0            15935190
#> 35   1934                  12736721                      0            17746308
#> 36   1935                  13332591                      0            21439629
#> 37   1936                  14524331                      0            25757225
#> 38   1937                  15194685                      0            27029929
#> 39   1938                  13705010                      0            23064881
#> 40   1939                  14375364                      0            26628560
#> 41   1940                  14375364                      0            28664617
#> 42   1941                  15343652                      0            33701276
#> 43   1942                  14896750                      0            33604721
#> 44   1943                  13705010                      0            32177661
#> 45   1944                  13481559                      0            31254927
#>     Sawnwood Imports Sawnwood Exports Roundwood consumed for lumber and panels
#> 1           436747.5        1382617.8                                 81168592
#> 2           488325.3        1283621.7                                 84738011
#> 3           613110.3        1198767.9                                 88378819
#> 4           539903.1        1574786.7                                 92233792
#> 5           524928.9        1628860.2                                 95589046
#> 6           666351.9        1482445.8                                 96945425
#> 7           839387.1        1736175.3                                102870661
#> 8           754533.3        1888413.0                                102942049
#> 9           639731.1        1504907.1                                 94089889
#> 10          824412.9        1638843.0                                 99800960
#> 11          790305.0        1948309.8                                 99943737
#> 12          698796.0        2324928.9                                 96517095
#> 13          853529.4        2599218.3                                100943174
#> 14          806943.0        2711751.3                                 98730134
#> 15          764516.1        1860258.1                                 91020189
#> 16          881814.0        1160250.5                                 83310244
#> 17         1020741.3        1115566.1                                 89592421
#> 18         1001607.6        1040371.2                                 80811650
#> 19         1007253.5        1003406.9                                 72102267
#> 20          958330.5        1421511.0                                 78170280
#> 21         1129842.6        1531235.4                                 78812775
#> 22          702423.4        1185421.6                                 65391759
#> 23         1312983.0        1748569.4                                 79526659
#> 24         1690708.8        2208035.5                                 92662122
#> 25         1492612.6        2457495.6                                 89306868
#> 26         1590685.1        2357735.0                                 92947675
#> 27         1640721.5        2529815.9                                 90306305
#> 28         1506663.2        2755218.2                                 85166342
#> 29         1268794.8        2938493.5                                 84024127
#> 30         1345060.6        2897412.3                                 88807149
#> 31         1033905.1        2163241.9                                 67319246
#> 32          635480.0        1580550.5                                 46116896
#> 33          324386.0        1081085.9                                 31696442
#> 34          312029.9        1211333.5                                 39834718
#> 35          249655.7        1262948.0                                 43618302
#> 36          393109.8        1244395.5                                 52970181
#> 37          596301.5        1234143.7                                 63678438
#> 38          628831.4        1390694.1                                 67105081
#> 39          475590.2         944562.7                                 57895979
#> 40          647842.6        1044268.3                                 66819527
#> 41          673576.6         891846.6                                 72530598
#> 42         1214960.2         648846.1                                 84880788
#> 43         1337610.1         432735.3                                 84952176
#> 44          778995.5         296040.6                                 80026378
#> 45          910680.5         347049.6                                 76885289
#>     Log Exports (tons) Imported logs for lumber and panels (1000 tons)
#> 1            200475.31                                            0.00
#> 2            214794.32                                            0.00
#> 3            200366.23                                            0.00
#> 4            227467.06                                            0.00
#> 5            243749.56                                            0.00
#> 6            221232.42                                            0.00
#> 7            261335.05                                            0.00
#> 8            282599.73                                            0.00
#> 9            222364.79                                            0.00
#> 10           232601.94                                            0.00
#> 11           259769.87                                            0.00
#> 12           294817.59                                            0.00
#> 13           312764.94                                            0.00
#> 14           320336.33                                            0.00
#> 15           203163.12                                            0.00
#> 16           147961.39                                            0.00
#> 17           157006.44                                            0.00
#> 18           127204.38                                            0.00
#> 19           107800.57                                            0.00
#> 20           161703.96                                            0.00
#> 21           159273.26                                            0.00
#> 22           128972.14                                            0.00
#> 23           186317.48                                            0.00
#> 24           246248.78                                            0.00
#> 25           287007.96                                            0.00
#> 26           251191.80                                            0.00
#> 27           274460.93                                            0.00
#> 28           286809.05                                            0.00
#> 29           298672.19                                            0.00
#> 30           317166.59                                            0.00
#> 31           232428.47                                            0.00
#> 32           160095.62                                            0.00
#> 33           107440.70                                            0.00
#> 34           126960.54                                            0.00
#> 35           136016.67                                            0.00
#> 36           153289.64                                            0.00
#> 37           159209.65                                            0.00
#> 38           174898.44                                            0.00
#> 39           116127.01                                            0.00
#> 40           140282.17                                            0.00
#> 41           127277.27                                            0.00
#> 42           123731.75                                            0.00
#> 43            88513.20                                            0.00
#> 44            93539.43                                            0.00
#> 45            74959.42                                            0.00
#>     SP.Production   SP.Imports SP.Exports NSP.Production  NSP.Imports
#> 1           0.000       0.0000      0.000       12557.07 0.000000e+00
#> 2           0.000       0.0000      0.000       12557.07 0.000000e+00
#> 3           0.000       0.0000      0.000       25114.13 0.000000e+00
#> 4           0.000       0.0000      0.000       37671.20 0.000000e+00
#> 5           0.000       0.0000      0.000       50228.26 0.000000e+00
#> 6           0.000       0.0000      0.000       87899.46 0.000000e+00
#> 7           0.000       0.0000      0.000      150684.78 0.000000e+00
#> 8           0.000       0.0000      0.000      163241.85 0.000000e+00
#> 9           0.000       0.0000      0.000      175798.91 0.000000e+00
#> 10          0.000       0.0000      0.000      200913.04 0.000000e+00
#> 11       3164.774       0.0000      0.000      226027.17 0.000000e+00
#> 12       6329.548       0.0000      0.000      200913.04 0.000000e+00
#> 13       9494.322       0.0000      0.000      200913.04 0.000000e+00
#> 14      12659.096       0.0000      0.000      200913.04 0.000000e+00
#> 15      15823.870       0.0000      0.000      213470.11 0.000000e+00
#> 16      18988.643       0.0000      0.000      213470.11 0.000000e+00
#> 17      22153.417       0.0000      0.000      227385.87 0.000000e+00
#> 18      25318.191       0.0000      0.000      228744.57 0.000000e+00
#> 19      28482.965       0.0000      0.000      242660.33 0.000000e+00
#> 20      31647.739       0.0000      0.000      269133.15 0.000000e+00
#> 21      34812.513       0.0000      0.000      207706.52 0.000000e+00
#> 22      41729.509       0.0000      0.000      196508.15 0.000000e+00
#> 23      48646.505       0.0000      0.000      235538.04 0.000000e+00
#> 24      55563.501       0.0000      0.000      299682.07 0.000000e+00
#> 25      62480.497       0.0000      0.000      301040.76 0.000000e+00
#> 26      69625.026       0.0000      0.000      389511.05 0.000000e+00
#> 27      79990.408       0.0000      0.000      433427.84 0.000000e+00
#> 28      90355.789       0.0000   1978.684      509759.84 9.114976e+01
#> 29     100721.171       0.0000   3957.368      519655.98 1.822995e+02
#> 30     162913.460       0.0000   5936.052      643797.34 2.734493e+02
#> 31     138794.986       0.0000   7914.735      511981.10 3.645990e+02
#> 32     106940.399       0.0000   9893.419      424763.35 4.557488e+02
#> 33     142207.978       0.0000  11872.103      362725.60 5.468986e+02
#> 34     177475.556       0.0000  13850.787      371324.21 6.380483e+02
#> 35     197953.505       0.0000  15829.471      393777.42 7.291981e+02
#> 36     218431.454       0.0000  17808.155      443299.52 8.366522e+02
#> 37     318545.871       0.0000  25881.397      511112.56 1.279087e+03
#> 38     329922.509       0.0000  20609.463      572987.92 4.051043e+03
#> 39     295792.594       0.0000      0.000      630904.83 2.559957e+03
#> 40     469627.626       0.0000      0.000      677935.99 3.186848e+03
#> 41     523325.359       0.0000      0.000      791271.26 4.203109e+03
#> 42     728104.847       0.0000      0.000     1286597.34 4.717870e+03
#> 43     837320.574       0.0000      0.000     1565222.58 1.162043e+03
#> 44     655294.362       0.0000      0.000     1563294.44 3.855870e+02
#> 45     655294.362       0.0000      0.000     1546259.60 4.332609e+01
#>     NSP.Exports Other Products Production Special Sawnwood Prod Special
#> 1        0.0000                          21749255              35073480
#> 2        0.0000                          22196158              36571607
#> 3        0.0000                          22717544              38057299
#> 4        0.0000                          23462381              39560611
#> 5        0.0000                          23834800              41059795
#> 6        0.0000                          24430670              41504455
#> 7        0.0000                          26814150              43878360
#> 8        0.0000                          31432142              43864469
#> 9        0.0000                          28750728              40016149
#> 10       0.0000                          30538338              42384900
#> 11       0.0000                          30910756              42317373
#> 12       0.0000                          30091435              40840912
#> 13       0.0000                          30016951              42673181
#> 14       0.0000                          29719016              41660878
#> 15       0.0000                          28676244              38245816
#> 16       0.0000                          28080374              34581031
#> 17      16.4721                          28750728              37379986
#> 18      32.9442                          28750728              33158298
#> 19      49.4163                          28601760              29656906
#> 20      65.8884                          28527276              32350761
#> 21      82.3605                          28154858              32844302
#> 22      98.8326                          25622410              26930970
#> 23     115.3047                          25249991              32535889
#> 24     131.7768                          25398959              38063581
#> 25     148.2489                          24654121              36917836
#> 26     855.3324                          22270641              38033494
#> 27    1173.9856                          22196158              36981363
#> 28    2268.9691                          21376836              34709352
#> 29    3514.8935                          20631999              34110933
#> 30    5632.9216                          20557515              36279786
#> 31    6392.4806                          17801616              27554992
#> 32    7269.4381                          14449848              18737224
#> 33    7509.0891                          12364302              12642006
#> 34    8520.2164                          12438786              15985979
#> 35    9766.1409                          12736721              17801647
#> 36   29689.4339                          13332591              21501673
#> 37   31840.3190                          14524331              25821623
#> 38   48364.9433                          15194685              27100378
#> 39   29231.1783                          13705010              23111144
#> 40   37005.8646                          14375364              26684464
#> 41   95830.8533                          14375364              28714918
#> 42  127211.0561                          15343652              33750403
#> 43  180739.4845                          14896750              33639735
#> 44  332147.2709                          13705010              32215272
#> 45  239074.8762                          13481559              31285399
#>     SP Prod Special NSP Prod Special Sawnwood Consumption SP.Consumption
#> 1             0.000         12588.08             34041196          0.000
#> 2             0.000         12588.89             35683844          0.000
#> 3             0.000         25171.07             37385555          0.000
#> 4             0.000         37764.10             38428403          0.000
#> 5             0.000         50356.34             39851428          0.000
#> 6             0.000         88100.05             40593863          0.000
#> 7             0.000        151067.59             42870384          0.000
#> 8             0.000        163689.98             42610501          0.000
#> 9             0.000        176214.38             39056625          0.000
#> 10            0.000        201381.30             41471915          0.000
#> 11         3173.000        226614.65             41049664       3164.774
#> 12         6348.882        201526.75             39090408       6329.548
#> 13         9523.739        201535.56             40795681       9494.322
#> 14        12700.169        201564.92             39621336      12659.096
#> 15        15859.189        213946.59             37064897      15823.870
#> 16        19022.368        213849.24             34241287      18988.643
#> 17        22192.240        227784.35             37219769      22153.417
#> 18        25358.044        229104.63             33067422      25318.191
#> 19        28525.550        243023.13             29616478      28482.965
#> 20        31713.206        269689.88             31820797      31647.739
#> 21        34882.866        208126.28             32376667      34812.513
#> 22        41811.812        196895.73             26394961      41729.509
#> 23        48760.476        236089.87             32024254      48646.505
#> 24        55711.161        300478.47             37445369      55563.501
#> 25        62681.293        302008.22             35834689      62480.497
#> 26        69813.188        390563.71             37163935      69625.026
#> 27        80233.516        434745.12             35980215      79990.408
#> 28        90660.074        511476.53             33344301      88377.105
#> 29       101079.194        521503.15             32320413      96763.803
#> 30       163495.290        646096.61             34598326     156977.408
#> 31       139274.194        513748.78             26330845     130880.251
#> 32       107311.645        426237.92             17727331      97046.980
#> 33       142690.017        363955.13             11842599     130335.875
#> 34       178041.204        372507.69             15035887     163624.769
#> 35       198570.792        395005.35             16733016     182124.034
#> 36       219063.570        444582.37             20588343     200623.299
#> 37       319342.303        512390.45             25119383     292664.474
#> 38       330782.398        574481.32             26268066     309313.046
#> 39       296385.891        632170.29             22595908     295792.594
#> 40       470613.571        679359.26             26232134     469627.626
#> 41       524243.694        792659.79             28446347     523325.359
#> 42       729166.214       1288472.83             34267390     728104.847
#> 43       838192.994       1566853.42             34509596     837320.574
#> 44       656060.308       1565121.71             32660616     655294.362
#> 45       655933.243       1547767.13             31818558     655294.362
#>     NSP.Consumption
#> 1          12557.07
#> 2          12557.07
#> 3          25114.13
#> 4          37671.20
#> 5          50228.26
#> 6          87899.46
#> 7         150684.78
#> 8         163241.85
#> 9         175798.91
#> 10        200913.04
#> 11        226027.17
#> 12        200913.04
#> 13        200913.04
#> 14        200913.04
#> 15        213470.11
#> 16        213470.11
#> 17        227369.40
#> 18        228711.62
#> 19        242610.91
#> 20        269067.26
#> 21        207624.16
#> 22        196409.32
#> 23        235422.74
#> 24        299550.29
#> 25        300892.51
#> 26        388655.72
#> 27        432253.85
#> 28        507582.02
#> 29        516323.38
#> 30        638437.87
#> 31        505953.22
#> 32        417949.66
#> 33        355763.41
#> 34        363442.05
#> 35        384740.47
#> 36        414446.74
#> 37        480551.33
#> 38        528674.02
#> 39        604233.61
#> 40        644116.97
#> 41        699643.51
#> 42       1164104.16
#> 43       1385645.14
#> 44       1231532.76
#> 45       1307228.05
<<<<<<< HEAD
#>  [ reached getOption("max.print") -- omitted 76 rows ]

>>>>>>> master