Estimates U.S. HWP contribution to annual greenhouse gas removals.

All units returned are in Thousand metric tons CO2 Equivalent

finalCarbonContribution(Years = 1990:2015, approach = c("Production",
  "Stock Change", "Atmospheric Flow"), decaytype = c("Exponential", "K=2"),
  plot = FALSE, halflives = halfLives)

Arguments

Years

years to return carbon totals for

approach

The approach used to calculate carbon contribution.

decaytype

Type of decay method to use

plot

whether to return a plot or not

halflives

data frame of half lives to use

Value

A vector of carbon contributions for all Years. Plot returned is optional

Details

The "Production" approach corresponds to `06 IPCC Tables`$R9 in the WOODCARB spreadsheet The "Stock Change" approach corresponds to `06 IPCC Tables`$P9 in the WOODCARB spreadsheet

The plot argument returns a simple plot of time vs. carbon contribution.

Examples

finalCarbonContribution()

#>  [1] -131771.55 -123758.01 -123790.61 -120707.52 -122498.31 -118411.13
#>  [7] -112219.10 -117343.83 -114188.45 -119182.37 -112968.80  -93478.73
#> [13]  -98188.13  -93967.40 -103967.46 -102683.17 -107666.14 -102762.70
#> [19]  -76830.26  -54360.98  -60549.58  -68943.22  -68312.97  -70814.78
#> [25]  -86538.26  -91250.87

finalCarbonContribution(approach = "Stock Change")

#>  [1] -129621.60 -116344.77 -119985.28 -126805.46 -129953.63 -125980.54
#>  [7] -122340.29 -131434.35 -137217.83 -147057.27 -141195.29 -125039.45
#> [13] -130714.01 -125812.07 -143193.24 -142102.08 -138121.66 -115131.55
#> [19]  -73079.21  -42305.07  -49151.35  -52374.83  -57982.74  -67784.59
#> [25] -110330.38 -115278.27

finalCarbonContribution(approach = "Production",
                         decaytype ="Exponential")

#>  [1] -131771.55 -123758.01 -123790.61 -120707.52 -122498.31 -118411.13
#>  [7] -112219.10 -117343.83 -114188.45 -119182.37 -112968.80  -93478.73
#> [13]  -98188.13  -93967.40 -103967.46 -102683.17 -107666.14 -102762.70
#> [19]  -76830.26  -54360.98  -60549.58  -68943.22  -68312.97  -70814.78
#> [25]  -86538.26  -91250.87