<<<<<<< HEAD

Estimated production, net imports, and apparent consumption of timber products in the US, 1900-1962

>>>>>>> master
Dataset containing estimated statistics for timber products. All vectors are numeric. Most data is for products in the Industrial Roundwood Class. Each column except the first two and the last one are in this category.

hair1963
<<<<<<< HEAD
Format

A data frame with 85 observations on 38 variables

=======

Format

A data frame with 85 observations on 38 variables

>>>>>>> master
Dom.Prod.Tot Total Domestic Production

ApparentConsumption Apparent Consumption

Dom.Prod Domestic Production, Total Ind RW

Imports Net Imports, Total Ind RW

Prod.SW Softwood Production, Total Ind RW

Prod.HW Hardwood Production , Total Ind RW

LogChipImports.SW Softwood Log and Chip Imports, Total Ind RW

LogChipExports.SW Softwood Log and Chip Exports, Total Ind RW

LogChipImports.HW Hardwood Log and Chip Imports, Total Ind RW

LogChipExports.HW Hardwood Log and Chip Exports, Total Ind RW

SW.PctOfProd Softwoods as Percent of Production, Total Ind RW

Exports.SW.PctOfProd Softwood Exports as Percent of Production, Total Ind RW

Exports.HW.PctOfProd Hardwood Exports as Percent of Production, Total Ind RW

Imports.SW.PctOfConsump Softwood Imports as Percent of Production, Total Ind RW

Imports.HW.PctOfConsump Hardwood Imports as Percent of Production, Total Ind RW

SW.Lumber.PctOfTotLumbProd Softwood Lumber as Percent of Total Lumber Production, Total Ind RW

DV.Exports.Log.PctOfProd Dollar Value of Log Exports as Percent of Production, Total Ind RW

DV.Imports.Log.PctOfConsump Dollar Value of Log Imports as Percent of Consumption, Total Ind RW

Estimated.Imports Estimated Imports, Total Ind RW

Estimated.Exports Estimated Exports, Total Ind RW

Adj.Imports Adjusted Imports, Total Ind RW

Adj.Exports Adjust Exports, Total Ind RW

Estimated.Imports(tons) Estimated Imports in tons, Total Ind RW

Estimated.Exports(tons) Estimated Exports in tons, Total Ind RW

Adj.Imports(tons) Adjusted Imports in tons, Total Ind RW

Adj.Exports(tons) Adjust Exports in tons, Total Ind RW

ApparentConsumption.Total Apparent Conusmption, Total Ind RW. Columns may not add to total because of rounding.

SawLogs.Dom.Prod Domestic Production of Saw Logs

SawLogs.NetImports Net Imports of Saw Logs. Net imports of lumber converted to cubic feet roundwood. Small quantities of imported saw logs (roundwood form) are included under domestic production.

SawLogs.ApparentConsumption Apparent Consumption of Saw Logs. Columns may not add to total because of rounding. Includes changes in stocks beginning in 1935.

VeneerLogs.Dom.Prod Domestic Production of Veneer Logs

VeneerLogs.NetImports Net Imports of Veneer Logs. Net imports of veneer logs represent the equivalent net imports of veneer and plywood converted to board feet log scale, and then to cubic feet round-wood. The small volume of veneer logs imported (roundwood form) is included under domestic production.

VeneerLogs.ApparentConsumption Apparent Consumption of Veneer Logs. Columns may not add to total because of rounding.

PulpWood.Dom.Prod Domestic Production of Pulpwood

PulpWood.NetImports Net Imports of Pulpwood

PulpWood.ApparentConsumption Apparent Consumption of Pulpwood

Other.ApparentConsumption Apparent Consumption of Other Products

FuelWood.ApparentConsumption Apparent Consumption of Fuelwood

<<<<<<< HEAD
Source

US Department of Commerce and Agriculture and USFS Estimations

Source

US Department of Commerce and Agriculture and USFS Estimations

>>>>>>> master