Plywood production, imports, exports, and consumption, by softwoods and hardwoods, 1965-1999

Dataset containing statisics for plywood products(3/8-in. basis). All units are in Million Square ft. unless otherwise specified.

Prod.Tot Total Plywood Production

Prod.SW Softwood Plywood Production

Prod.HW Hardwood Plywood Production

Imports.Tot Total Plywood Imports

Imports.SW Softwood Plywood Imports

Imports.HW Hardwood Plywood Imports

Consump.Tot Total Plywood Consumption

Consump.SW Softwood Plywood Consumption

Consump.HW. Hardwood Plywood Consumption

PerCapConsump.Tot Total Per Capita Consumption, in Square ft.

PerCapConsump.SW Softwood Per Capita Consumption, in Square ft.

PerCapConsump.HW Hardwood Per Capita Consumption, in Square ft.

howard37

Format

A data frame with 56 observations of 15 variables

Source

US Dept. of Commerce, Bureau of Census, American Plywood Association The Engineered Wood Association, US Dept. of Agriculture, Foreign Agriculture Service