Paper and board new supply and recyclable paper consumption, exports, imports and total recovered, 1965-1999

Dataset containing consumption and production statistics for various board and paper products. All units are in thousand tons unless otherwise specified. Recovery rate is the ratio of total recovered paper collected to new supply of paper and paperboard.

howard47

Format

A data frame with 56 observations of 19 variables

Source

1972-Present, American Forest and Paper Association, 1999 Annual Statistical Summary Recovered Paper Utilization, p.81 (6). 1965-1971, American Forest and Paper Assocation (2,4,5).

Details

PaperBoardNewSupply New supply of Paper and Board

RecPap.ConsumedatPaperBoardMills Recyclable Paper Consumed at Paper and Board Mills

RecPap.MoldedPulpInsulationandOther Recyclable Paper for Molded pulp, insulation and other uses

RecPap.Exports Recyclable Paper Exports

RecPap.Imports Recyclable Paper Imports

RecPap.TotRecovered Total Recyclable Paper Recovered

RecPap.RecoveryRate Recyclable Paper Recovery Rate, as a percent.

NoName Recylable Paper Totals for certain years

Ratio.ExportstoProduction Ratio of Exports to Production

Ratio.ImportstoProduction Ratio of Imports to Production