Particleboard and medium-density fiberboard production, imports,

Dataset containing various statistics for particleboard and medium-density fiberboard. All units are in Million Square ft unless otherwise specified.

howard53

Format

A data frame with 56 observations of 7 variables

Source

National Particleboard Association (20); U.S. International Trade Commission (69); U.S. Department of Agriculture (32). Data may not add to totals because of rounding.

Details

Prod.Tot Total Particleboard and Medium-Density Fiberboard Production

Prod.ParticleBoard Particleboard Production

Prod.MediumDensityFiberboard Medium-Density Fiberboard Production

Imports Particleboard and Medium-Density Fiberboard Imports

Exports Particleboard and Medium-Density Fiberboard Exports

Consump.Tot Total Particleboard and Medium-Density Fiberboard Consumption

Consump.PerCap(Sq ft) Per Capita Consumption, in Square ft.