<<<<<<< HEAD

Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.

>>>>>>> master
Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.

howard6
<<<<<<< HEAD
Format

A data frame with 56 observations on 34 variables. (all units are in million cubic feet except when specified otherwise) =======

Format

A data frame with 56 observations on 34 variables. (all units are in million cubic feet except when specified otherwise) >>>>>>> master

AllProduction.Prod (All products production)

AllProduct.Consump(All products consumption)

Ind.RW.Tot.Prod(Industrial roundwood uses total production)

Ind.RW.Tot.Imports(Industrial roundwood uses total imports)

Ind.RW.Tot.Exports(Industrial roundwood uses total exports)

Ind.RW.Tot.Consump(Industrial roundwood uses total consumption)

Ind.RW.Lum.Prod(Industrial roundwood uses lumber production)

Ind.RW.Lum.Imports(Industrial roundwood uses lumber imports)

Ind.RW.Lum.Exports(Industrial roundwood uses lumber exports)

Ind.RW.Lum.Consump(Industrial roundwood uses lumber consumption)

Ind.RW.PlyandVen.Prod(Industrial roundwood uses plywood and veneer production)

Ind.RW.PlyamdVen.Imports(Industrial roundwood uses plywood and veneer imports)

Ind.RW.PlyandVen.Exports(Industrial roundwood uses plywood and veneer exports)

Ind.RW.PlyandVen.Consump(Industrial roundwood uses plywood and veneer consumption)

IndRW.Pulp.Prod(Industrial roundwood uses pulp-based products production)

Ind.RW.Pulp.Imports(Industrial roundwood uses pulp-based products imports)

Ind.RW.Pulp.Exports(Industrial roundwood uses pulp-based products exports)

Ind.RW.Pulp.Consump(Industrial roundwood uses pulp-based products consumption)

Ind.RW.OtherIndustrial.ProdAndConsump(Other industrial product production and consumption)

Ind.RW.Logs.Imports(Industrial roundwood uses log imports)

Ind.RW.Logs.Exports(Industrial roundwood uses log exports)

Ind.RW.Pulp.Imports(Industrial roundwood uses pulpwood chip imports)

Ind.RW.Pulp.Exports(Industrial roundwood uses pulpwood chip exports)

FuelWood.ProdAndConsumption(Fuelwood production and consumption)

UnNamed1

UnNamed2

UnNamed3

UnNamed4

UnNamed5

UnNamed6

UnNamed7

UnNamed8

UnNamed9

UnNamed10

<<<<<<< HEAD
Source

USFS Estimations and Bureau of Census Data

Source

USFS Estimations and Bureau of Census Data

>>>>>>> master