U.S. Annual industrial wood product in conventional units of measure.

"Lumber made at pallet plants" is a Forest Service estimate of lumber cut from roundwood at pallet plants. Other lumber (e.g., purchased lumber) used by pallet makers is accounted for under hardwood and softwood lumber production. Paper and paperboard production includes "Total Paper", "Total Paperboard", and "Building Paper" production. Total production of pulp, paper and paperboard includes market pulp produced for export. Fuelwood estimates are shown for comparison only and are not included in industrial wood products.

ince1

Format

A data frame with 151 observations of 25 variables:

SW.Ply

Softwood Plywood units: million ft^2, 3/8 in.

OSB.Wafer.board

OSB/ Wafer board units: million ft^2, 3/8 in.

Lam.Ven.Lumb

Laminated Veneer lumber units: million ft^3

HW.Ply.Ven

Hardwood, Plywood and veneer units: million ft^2, 3/8 in.

SW.Lumb

Softwood lumber units: million bd. ft.

HW.Lumb

Hardwood lumber units: million bd. ft.

Lumb.Pallet.Plant

Lumber made at Pallet plants units: million bd. ft.

PartBoard.Prod

Particleboard production units: million ft^2, 3/4 in.

Hardboard.Prod

Hardboard production units: million ft^2, 1/8 in.

MDF.Prod

MDF production units: million ft^2, 3/4 in.

Pulp.Paper.Board

Pulp, paper and board production units: Thousand tons

Other.Products

Other industrial products units: million ft^3

InsulatingBoard

Insulating board units: million ft^2, 1/2 in.

Fuelwood

Fuelwood units: million ft^3

Fuelwood.Tons

Fuelwood units: Thousand short tons

Log.Chip.Exports

Log and Chip exports units: Thousand short tons

RW.Dom.Prod

Roundwood Equivalent of Domestic Product Output units: Thousand short tons

Estimated.Tot.Dom.Timber

Estimated Total Domestic Timber Harvest units: Thousand short tons

Source

Softwood plywood, OSB, and LVL production data (1960-1998) from APA--The Engineered Wood Association (formerly American Plywood Assoc.); see APA Economics Report E-65 (1999).

Hardwood plywood production data from Dept. of Commerce (to 1988); 1989-1990 data from Hardwood Plywood & Veneer Association; later estimates based on trends in value of shipments (Dept. of Commerce); hardwood veneer based on Census of Manufactures data and trend in value of shipments.

Lumber production 1900-1946 based on H.B. Steer (USDA Forest Service, 1948, Table 1, "Estimated Production")

Lumber production 1947-1976 based on Commerce Department data and Forest Service estimates. Lumber production 1976-1998 from American Forest & Paper Association (AF&PA) Statistical Roundup report (1996 hardwood estimated by Forest Service; 1997-1998 hardwood estimate from Miller Freeman).

[Note that Commerce Dept. reported hardwood lumber production is understood to underestimate actual production as reflected in Forest Service estimates since 1900.]

Particleboard and MDF production data from Composite Panel Association (based on production data 1959-1977, and 1995-1997; otherwise based on shipments; 1998 data are estimated).

Hardboard production based on shipments data from American Hardboard Association (1964-1997) and as reported by Miller Freeman (1998); 1998 figure is estimated.

Pulp, paper, and paperboard production data are from AF&PA, formerly API (Statistics of Paper, Paperboard & Woodpulp).

Miscellaneous wood product production data through 1988 were obtained from U.S. Timber Production, Trade, Consumption, and Price Statistics reports (USDA Forest Service). Miscellaneous wood product production for 1996 based on timber product output tables in (draft) 1997 RPA Inventory Data Tables; intervening and subsequent data are extrapolated.

Insulating board production data are derived using earlier data from Commerce Department MA26A reports (in square feet of product output) to convert tonnage reported in AF&PA and earlier API reports (Statistics of Paper, Paperboard & Woodpulp).

Fuelwood data are Forest Service estimates (Skog and others, FPL).

Details

Units are in Million feet squared 3/8in.