Calculate total carbon in solidwood products.

Calculates total carbon and carbon by end use in solidwood products. Capable of calculating carbon totals with different approaches, halflives, and decay distributions.

swpcarbontotal(Yrs = 1990:2015, decaydistribution = c("Exponential", "K=2"),
  onlytotal = TRUE, lumberpre = TRUE, approach = c("Production",
  "Stock Change"), halflives = halfLives, fsp = fracstrpanels,
  fnsp = fracnonstrpanels, fsawn = fracsawnwood)

Arguments

Yrs

years to calculate totals for.

decaydistribution

type of decay distribution.

onlytotal

if false, total carbon is returned for each one of the 13 end uses. if true, one total is summed and returned

lumberpre

if false, carbon from lumber before 1900 is excluded from totals

approach

approach type to use to calculate. If onlytotal is false then the data returned won't correspond to data in WOODCARB II If onlytotal is true: For production approach, the data returned corresponds to the `SWCalcP$H4389` column in WOODCARB II. For stock change approach, the data returned corresponds to the `SWCalc$H4389` column in WOODCARB II.

halflives

data frame of half lives for end uses. Default half lives are used but a data frame with 13 columns with half lives for the appropriate years can be used.

fsp

Fraction of structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fnsp

Fraction of non-structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fsawn

Fraction of sawnwood products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

Value

Output is either a matrix with 13 columns or 1 column based on the onlytotal argument

Details

All units returned are in od tons of wood fiber.

Examples

swpcarbontotal(c(1990,2000,2010), onlytotal=FALSE)

#>   Years       EU1       EU2      EU3       EU4       EU5      EU6     EU7
#> 1  1990 706414309 182866038 63065693 258801895 116211294 32956528 3874448
#> 2  2000 836196543 176074498 74824810 316329446 138929063 32171235 4066189
#> 3  2010 940141701 168801706 82205622 342217459 154428853 30651352 4137418
#>         EU8      EU9      EU10      EU11      EU12      EU13 LumberPre1900
#> 1  97501711 32435774  83160265 180886699 231453156 478162361      70903425
#> 2 120999083 52418301 102787989 196832870 245974176 447244425      55508766
#> 3 138693102 68985296 116424536 206107483 256716678 411340193      43456619
#>   Total Carbon
#> 1   2538693597
#> 2   2800357395
#> 3   2964308018

swpcarbontotal(1950:1960, approach = "Stock Change")

#>  [1] 1818871465 1833139774 1846854799 1860447189 1873679776 1888553014
#>  [7] 1903747351 1913075329 1923424778 1937991525 1948188912