Lumber production, imports, exports, and consumption, by softwoods and hardwoods, 1950-87

Dataset containing various statistics for lumber products. All units are in Billion Board Ft unless otherwise specified. Year 1987 includes small volumes of mixed species (not classified as softwoods or hardwoods). Data may not add to totals because of rounding. Estimates of softwood lumber production for certain years in the late 1970's and early 1980's by the Bureau of the Census and some major softwood industry associations currently vary by as much as 5 percent. Cooperative studies are underway to resolve these differences. Revisions made by Census as a result of these studies will be reported in future issues of U.S. timber production, trade, consumption, and price statistics.

ulrich29

Format

A data frame with 38 observations of 15 variables

Source

U.S. Department of Commerce. Bureau of the Census (43.51.52) U.S. Department of Agriculture, Forest Service; National Forest Products Association (16). Data for selected years 1899.1949 in "The demand and price situation for forest products", 1964, table 7 (31)

Details

Prod.Tot Total Lumber Production

Prod.SW Softwood Lumber Production

Prod.HW Hardwood Lumber Prodcution

Imports.Tot Total Lumber Imports

Imports.SW Softwood Lumber Imports

Imports.HW Hardwood Lumber Imports

Exports.Tot Total Lumber Exports

Exports.SW Softwood Lumber Exports

Exports.HW Hardwood Lumber Exports

Consump.Tot Total Lumber Consumption

Consump.SW Softwood Lumber Consumption

Consump.HW Hardwood Lumber Consumption

PerCapitaConsump.Tot Total Lumber Consumption Per Capita, in Board ft.

PerCapitaConsump.SW Softwood Lumber Consumption Per Capita, in Board ft.

PerCapitaConsump.HW Hardwood Lumber Consumption Per Capita, in Board ft.