Plywood production, imports, exports, and consumption, by softwoods and hardwoods, 1950-87

Dataset containing various statistics for plywood products. (3/8-inch basis) All units are in Million Square ft unless otherwise specified. Year 1987 includes small volumes of mixed species (not classified as softwoods or hardwoods).

ulrich36

Format

A data frame with 38 observations of 15 variables:

Prod.Tot

Total Plywood Production

Prod.SW

Softwood Plywood Production

Prod.HW

Hardwood Plywood Production

Imports.Tot

Total Plywood Imports

Imports.SW

Softwood Plywood Imports

Imports.HW

Hardwood Plywood Imports

Exports.Tot

Total Plywood Exports

Exports.SW

Softwood Plywood Exports

Exports.HW

Hardwood Plywood Exports

Consump.Tot

Total Plywood Consumption

Consump.SW

Softwood Plywood Consumption

Consump.HW

Hardwood Plywood Consumption

PerCapitaConsump.Tot

Total Plywood Consumption Per Capita, in Square ft.

PerCapitaConsump.SW

Softwood Plywood Consumption Per Capita, in Square ft.

PerCapitaConsump.HW

Hardwood Plywood Consumption Per Capita, in Square ft.

U.S. Department of Commerce, Bureauot Census (40,50.51,52) American Plywood Association (5).