Production, imports, exports. and consumption of timber products, by major product, 1950-87

Dataset containing statistics for various timber products (hardwood and softwood). All units are Million cubic ft (roundwood equivalent) unless otherwise specified.

ulrich4

Format

A data frame with 38 observations of 24 variables

Source

Based on data published by the U.S. Departments of Commerce and Agriculture. Data for 1900-49 in "The demand and price situation for forest products 1964", table 2 (31),

Details

AllProducts.Prod Production for all timber products

AllProducts.Consump Consumption for all timber products

Tot.Prod Total Production for major timber products

Tot.Imports Total Imports for major timber products

Tot.Exports Total Exports for major timber products

Tot.Consump Total Consumption for major timber products

Lumber.Prod Lumber Products Prodcution

Lumber.Imports Lumber Products Imports

Lumber.Exports Lumber Products Exports

Lumber.Consump Lumber Products Consumption

PlywoodVeneer.Prod Plywood and Veneer Products Production

PlywoodVeneer.Imports Plywood and Veneer Products Imports

PlywoodVeneer.Exports Plywood and Veneer Products Exports

PlywoodVeneer.Consump Plywood and Veneer Products Consumption

PulpProducts.Prod Pulp Products Production

PulpProducts.Imports Pulp Products Imports

PulpProducts.Exports Pulp Products Exports

PulpProducts.Consump Pulp Products Consumption

OtherIndProducts.ProdandConsump Production and Consumption for other industrial products

Logs.Imports Log Imports

Logs.Exports Log Exports

PulpwoodChip.Exports Pulpwood-chip Exports

Fuelwood.ProdandConsump Production and Consumption for fuelwood products

LogChipExports.PercofProduction Log and Chip exports as a percent of production