<<<<<<< HEAD

Production, imports, exports, and consumption of hardwood timber products, by major product, 1950-87

>>>>>>> master
Dataset containing statistics for hardwood timber products. All units are in million cubic ft.

ulrich6
<<<<<<< HEAD
Format

A data frame with 48 observations of 22 variables.

Source

Based on data published by the U.S. Departments of Commerce and Agriculture.

Details

=======

Format

A data frame with 48 observations of 22 variables.

Source

Based on data published by the U.S. Departments of Commerce and Agriculture.

Details

>>>>>>> master
AllProduction.Prod Total production of all products

AllProduct.Consump Total consumption of all products

Indu.RW.Tot.Prod Total production of all industrial roundwood products

Indu.RW.Tot.Imports Total imports of all industrial roundwood products

Indu.RW.Tot.Exports Total exports of all industrial roundwood products

Indu.RW.Tot.Consump Total consumption of all industrial roundwood products

Indu.RW.Lum.Prod Production of lumber products

Indu.RW.Lum.Imports Imports of lumber products

Indu.RW.Lum.Exports Exports of lumber products

Indu.RW.Lum.Consump Consumption of lumber products

Indu.RW.PlyandVen.Prod Production of plywood and veneer products

Indu.RW.PlyandVen.Imports Imports of plywood and veneer products

Indu.RW.PlyandVen.Exports Exports of plywood and veneer products

Indu.RW.PlyandVen.Consump Consumption of plywood and veneer products

Indu.RW.Pulp.Prod Production of pulp products

Indu.RW.Pulp.Imports Imports of pulp products

Indu.RW.Pulp.Exports Exports of pulp products

Indu.RW.Pulp.Consump Consumption of pulp products

Indu.RW.OtherIndustrial.ProdAndConsump Production and consumption of Other industrial roundwood products

Indu.RW.Logs.Imports Imports of logs

Indu.RW.Logs.Exports Exports of logs

FuelWood.ProdAndConsumption Production and Consumption of Fuelwood

UnNamed1 Unnamed

<<<<<<< HEAD

>>>>>>> master