Calculate carbon in SWDS for the Production Method

Column names with only letters come from Calc sheet, otherwise sheet is specified (need to change names)

calculatedumpcarbonproduction(Years, totalcarbonstockchange = FALSE)

Arguments

Years

years to calculate

totalcarbonstockchange

If true, returns carbon stock change, or returns data frame of intermediate calculations

Value

Either a data frame of values from Calculation spreadsheet or total carbon output

Examples

calculatedumpcarbonproduction(1995:2020)

#>     Years       CA       CB ParamResults_V ParamResults_P       CI       CS
#> 96   1995 112.0782 35.31055      0.3930708      0.3930708 13.87955 6.107000
#> 97   1996 114.9872 36.22701      0.3673884      0.3673884 13.30938 5.856129
#> 98   1997 117.1731 36.91571      0.3822584      0.3822584 14.11134 6.208990
#> 99   1998 120.0203 37.81273      0.3898972      0.3898972 14.74308 6.486954
#> 100  1999 121.8382 38.38547      0.3902860      0.3902860 14.98131 6.591777
#> 101  2000 124.0266 39.07491      0.3743003      0.3743003 14.62575 6.435330
#> 102  2001 124.3738 39.18430      0.3447074      0.3447074 13.50712 5.943131
#> 103  2002 123.1034 38.78405      0.3494005      0.3494005 13.55117 5.962513
#> 104  2003 122.8649 38.70892      0.3231569      0.3231569 12.50905 5.503984
#> 105  2004 122.5709 38.61629      0.3156480      0.3156480 12.18916 5.363228
#> 106  2005 123.2716 38.83705      0.3079519      0.3079519 11.95994 5.262375
#> 107  2006 124.6243 39.26323      0.3244935      0.3244935 12.74066 5.605892
#> 108  2007 126.9862 40.00735      0.3203581      0.3203581 12.81668 5.639339
#> 109  2008 129.1539 40.69028      0.3162227      0.3162227 12.86719 5.661563
#> 110  2009 130.0205 40.96332      0.3120873      0.3120873 12.78413 5.625018
#> 111  2010 128.1300 40.36769      0.3079519      0.3079519 12.43131 5.469775
#> 112  2011 128.8919 40.60774      0.3079519      0.3079519 12.50523 5.502301
#> 113  2012 128.8901 40.60716      0.3079519      0.3079519 12.50505 5.502222
#> 114  2013 128.0167 40.33202      0.3079519      0.3079519 12.42032 5.464941
#> 115  2014 125.8527 39.65024      0.3079519      0.3079519 12.21037 5.372561
#> 116  2015 128.5943 40.51400      0.3079519      0.3079519 12.47636 5.489599
#> 117  2016 131.1253 41.31137      0.3079519      0.3079519 12.72191 5.597642
#> 118  2017 133.1799 41.95869      0.3079519      0.3079519 12.92126 5.685353
#> 119  2018 135.1429 42.57713      0.3079519      0.3079519 13.11171 5.769151
#> 120  2019 136.9671 43.15185      0.3079519      0.3079519 13.28869 5.847025
#> 121  2020 138.7210 43.70442      0.3079519      0.3079519 13.45886 5.921897
#>           CU       CV        CX ParamResults_Q   Dumps_S ParamResults_O
#> 96  7.772546 105.3090 2.7384315     0.01613588 0.5697666      0.7274305
#> 97  7.453256 107.6178 2.3087742     0.01596480 0.5783569      0.7083651
#> 98  7.902351 110.2499 2.6320488     0.01575570 0.5816329      0.6882236
#> 99  8.256124 113.0995 2.8496018     0.01570413 0.5938160      0.6754670
#> 100 8.389535 115.9464 2.8469205     0.01599190 0.6138567      0.6774638
#> 101 8.190420 118.4734 2.5270072     0.01611950 0.6298679      0.6727094
#> 102 7.563984 120.2872 1.8138626     0.01612050 0.6316706      0.6727515
#> 103 7.588653 122.0419 1.7546526     0.01622322 0.6292020      0.6770379
#> 104 7.005070 123.1595 1.1176434     0.01606186 0.6217373      0.6703042
#> 105 6.825927 124.0552 0.8956834     0.01603494 0.6192097      0.6691806
#> 106 6.697568 124.7875 0.7323176     0.01605844 0.6236624      0.6701613
#> 107 7.134772 125.9037 1.1161648     0.01605844 0.6305062      0.6701613
#> 108 7.177341 127.0096 1.1058695     0.01605844 0.6424557      0.6701613
#> 109 7.205625 128.0920 1.0824114     0.01605844 0.6534223      0.6701613
#> 110 7.159114 129.0806 0.9886399     0.01605844 0.6578070      0.6701613
#> 111 6.961532 129.8356 0.7549676     0.01605844 0.6482421      0.6701613
#> 112 7.002928 130.5956 0.7600320     0.01605844 0.6520968      0.6701613
#> 113 7.002829 131.3209 0.7252627     0.01605844 0.6520876      0.6701613
#> 114 6.955380 131.9678 0.6468904     0.01605844 0.6476693      0.6701613
#> 115 6.837805 132.4729 0.5051667     0.01605844 0.6367209      0.6701613
#> 116 6.986762 133.0972 0.6242815     0.01605844 0.6505915      0.6701613
#> 117 7.124271 133.8243 0.7270360     0.01605844 0.6633961      0.6701613
#> 118 7.235904 134.6247 0.8004069     0.01605844 0.6737911      0.6701613
#> 119 7.342555 135.4903 0.8656758     0.01605844 0.6837222      0.6701613
#> 120 7.441668 136.4111 0.9207721     0.01605844 0.6929513      0.6701613
#> 121 7.536960 137.3808 0.9697092     0.01605844 0.7018247      0.6701613
#>      Dumps_T    Dumps_V TotalCarbonOutput       CM       CN   Dumps_N   Dumps_O
#> 96  36.21400 -2.4728557          50210416 3.810474 58.61930 0.3674957 105.42418
#> 97  33.94073 -2.2732702          50653218 3.743328 60.54586 0.3668059 101.52599
#> 98  31.84668 -2.0940561          51117344 3.670215 62.34532 0.3653187  97.78353
#> 99  29.92616 -1.9205154          51432516 3.624396 64.04787 0.3663680  94.19295
#> 100 28.17299 -1.7531767          51997257 3.675025 65.74998 0.3771781  90.75751
#> 101 26.57046 -1.6025248          52349272 3.673938 67.40145 0.3827613  87.46592
#> 102 25.09380 -1.4766566          52340157 3.673528 69.00440 0.3827185  84.30699
#> 103 23.72932 -1.3644842          52810146 3.730131 70.61561 0.3886156  81.28108
#> 104 22.46370 -1.2656146          52987484 3.705433 72.15591 0.3860424  78.37468
#> 105 21.29385 -1.1698536          53631085 3.744153 73.68893 0.3900764  75.58933
#> 106 20.21878 -1.0750754          54016791 3.776608 75.20879 0.3934576  72.91952
#> 107 19.23334 -0.9854373          54207510 3.789942 76.69732 0.3948468  70.35868
#> 108 18.33530 -0.8980368          54099798 3.782411 78.13518 0.3940622  67.90032
#> 109 17.51698 -0.8183169          53062688 3.709901 79.46075 0.3865079  65.53382
#> 110 16.76614 -0.7508476          52321857 3.658105 80.69742 0.3811117  63.25755
#> 111 16.06466 -0.7014760          52298731 3.656489 81.89649 0.3809433  61.07289
#> 112 15.42111 -0.6435515          52528806 3.672574 83.07625 0.3826192  58.97791
#> 113 14.82743 -0.5936809          52614493 3.678565 84.22746 0.3832433  56.96799
#> 114 14.27569 -0.5517430          52920700 3.699974 85.36591 0.3854737  55.04124
#> 115 13.75661 -0.5190794          53140912 3.715370 86.48614 0.3870777  53.19371
#> 116 13.29055 -0.4660521          53449698 3.736959 87.59470 0.3893269  51.42282
#> 117 12.87244 -0.4181182          53795923 3.761165 88.69447 0.3918488  49.72574
#> 118 12.49631 -0.3761222          54178838 3.787937 89.78819 0.3946380  48.09976
#> 119 12.15850 -0.3378091          54597712 3.817223 90.87848 0.3976890  46.54226
#> 120 11.85539 -0.3031131          55051829 3.848973 91.96783 0.4009968  45.05072
#> 121 11.58396 -0.2714344          55513960 3.881283 93.05682 0.4043630  43.62255
#>           BX       CH       CK       CL       CT C_SWP_StockChange_LFDumps
#> 96  22.77507 16.56728 12.75680 285.5602 137.6788                  10.74707
#> 97  22.97592 16.27534 12.53201 298.0922 143.5349                  10.56039
#> 98  23.18644 15.95746 12.28724 310.3795 149.7439                  10.34424
#> 99  23.32940 15.75824 12.13385 322.5133 156.2309                  10.24582
#> 100 23.58557 15.97837 12.30334 334.8167 162.8227                  10.57000
#> 101 23.74524 15.97364 12.29971 347.1164 169.2580                  10.65958
#> 102 23.74110 15.97186 12.29833 359.4147 175.2011                  10.74236
#> 103 23.95429 16.21796 12.48783 371.9025 181.1636                  11.07313
#> 104 24.03473 16.11058 12.40514 384.3077 186.6676                  11.03905
#> 105 24.32666 16.27893 12.53477 396.8425 192.0309                  11.28244
#> 106 24.50161 16.42003 12.64343 409.4859 197.2932                  11.49348
#> 107 24.58812 16.47801 12.68807 422.1740 202.8991                  11.61575
#> 108 24.53926 16.44527 12.66285 434.8368 208.5385                  11.64236
#> 109 24.06884 16.13000 12.42010 447.2569 214.2000                  11.37918
#> 110 23.73280 15.90481 12.24670 459.5036 219.8250                  11.20710
#> 111 23.72231 15.89778 12.24129 471.7449 225.2948                  11.25570
#> 112 23.82667 15.96771 12.29514 484.0400 230.7971                  11.37992
#> 113 23.86554 15.99376 12.31520 496.3552 236.2993                  11.45648
#> 114 24.00443 16.08684 12.38687 508.7421 241.7643                  11.59857
#> 115 24.10432 16.15378 12.43841 521.1805 247.1368                  11.71112
#> 116 24.24438 16.24765 12.51069 533.6912 252.6264                  11.84836
#> 117 24.40143 16.35289 12.59173 546.2829 258.2241                  11.99442
#> 118 24.57511 16.46929 12.68135 558.9643 263.9094                  12.14909
#> 119 24.76511 16.59662 12.77940 571.7437 269.6786                  12.31219
#> 120 24.97110 16.73466 12.88569 584.6294 275.5256                  12.48351
#> 121 25.18072 16.87514 12.99386 597.6232 281.4475                  12.65467
#>     C_PAPER_StockChange_LFDumps TotalC_StockChange_SWDSProduction
#> 96                     6.372576                          17.11964
#> 97                     5.891633                          16.45202
#> 98                     6.746982                          17.09122
#> 99                     7.416041                          17.66186
#> 100                    7.685521                          18.25552
#> 101                    7.359813                          18.01940
#> 102                    6.280337                          17.02270
#> 103                    6.352681                          17.42581
#> 104                    5.356013                          16.39506
#> 105                    5.089058                          16.37150
#> 106                    4.919617                          16.41309
#> 107                    5.736620                          17.35237
#> 108                    5.847172                          17.48953
#> 109                    5.925657                          17.30484
#> 110                    5.862811                          17.06991
#> 111                    5.523266                          16.77896
#> 112                    5.618781                          16.99870
#> 113                    5.633804                          17.09029
#> 114                    5.560089                          17.15866
#> 115                    5.358648                          17.06976
#> 116                    5.647828                          17.49619
#> 117                    5.906560                          17.90098
#> 118                    6.109638                          18.25873
#> 119                    6.297017                          18.60920
#> 120                    6.464684                          18.94819
#> 121                    6.620172                          19.27484

calculatedumpcarbonproduction(1990:1991, totalcarbonstockchange = TRUE)

#> [1] 18.13123 18.72634