<<<<<<< HEAD

Calculate carbon in SWDS for the Stock Change Method

>>>>>>> master
Calculate carbon in SWDS for the Stock Change Method

calculatedumpcarbonstockchange(Years, totalcarbonstockchange = FALSE)
<<<<<<< HEAD
Arguments
=======
Arguments
>>>>>>> master

Years

years to calculate

totalcarbonstockchange

Whether to return data frame of intermedate calculations or just totals

<<<<<<< HEAD
Value

If totalcarbonstockchange is false: a data frame of intermediate calculations. If totalcarbonstockchange is true: a numeric vector of stock change for total carbon from dumps.

Examples
=======

Value

If totalcarbonstockchange is false: a data frame of intermediate calculations. If totalcarbonstockchange is true: a numeric vector of stock change for total carbon from dumps.

Examples
>>>>>>> master

calculatedumpcarbonstockchange(Years = 1990:2015)

#>     Years TotalCarbonOutputStockChange ParamResults_Q   Dumps_C   Dumps_D
#> 91   1990                     50789440     0.01517695 0.3496421 129.91531
#> 92   1991                     50653922     0.01638197 0.3763961 125.03896
#> 93   1992                     51521203     0.01490181 0.3482497 120.33218
#> 94   1993                     52205309     0.01503865 0.3561142 115.82271
#> 95   1994                     52891673     0.01517198 0.3639950 111.50260
#> 96   1995                     53352870     0.01613588 0.3904956 107.38209
#> 97   1996                     54077914     0.01596480 0.3916059 103.42877
#> 98   1997                     54787478     0.01575570 0.3915479  99.63476
#> 99   1998                     55587859     0.01570413 0.3959676  95.99796
#> 100  1999                     56476871     0.01599190 0.4096724  92.52093
#> 101  2000                     57095559     0.01611950 0.4174646  89.19155
#> 102  2001                     57475588     0.01612050 0.4202695  85.99909
#> 103  2002                     58306899     0.01622322 0.4290647  82.94378
#> 104  2003                     58878411     0.01606186 0.4289610  80.01154
#> 105  2004                     60241448     0.01603494 0.4381558  77.20634
#> 106  2005                     61080815     0.01605844 0.4449119  74.52072
#> 107  2006                     61324267     0.01605844 0.4466852  71.94507
#> 108  2007                     60853367     0.01605844 0.4432552  69.46997
#> 109  2008                     59250726     0.01605844 0.4315815  67.08345
#> 110  2009                     58232243     0.01605844 0.4241629  64.78602
#> 111  2010                     58163727     0.01605844 0.4236639  62.58073
#> 112  2011                     58298544     0.01605844 0.4246459  60.46530
#> 113  2012                     58507421     0.01605844 0.4261673  58.43661
#> 114  2013                     58959537     0.01605844 0.4294605  56.49286
#> 115  2014                     59492926     0.01605844 0.4333457  54.63121
#> 116  2015                     60054468     0.01605844 0.4374360  52.84854
#>            Q   Dumps_F        X        Y       AA        V        D         J
<<<<<<< HEAD
#> 91  17.13246 -5.107960 3.940466 49.98057 2.440731 13.19200 32.09376  94.96866
#> 92  18.12870 -4.876355 4.169601 52.57266 2.592087 13.95910 32.25799  96.74643
#> 93  16.49147 -4.706775 3.793037 54.72364 2.150980 12.69843 32.36923  98.18288
#> 94  16.58525 -4.509473 3.814608 56.83290 2.109259 12.77064 33.88667 100.42909
#> 95  16.67655 -4.320109 3.835608 58.90109 2.068199 12.84095 34.95360 102.94850
#> 96  17.60415 -4.120510 4.048955 61.11618 2.215081 13.55520 36.44439 105.99795
#> 97  17.37573 -3.953325 3.996417 63.21572 2.099543 13.37931 36.06403 108.02760
#> 98  17.10318 -3.794001 3.933730 65.19324 1.977518 13.16944 36.48534 109.89137
#> 99  17.03138 -3.636803 3.917218 67.09711 1.903878 13.11417 38.21951 112.62733
#> 100 17.35492 -3.477032 3.991631 69.01777 1.920659 13.36329 39.06623 115.35170
#> 101 17.42191 -3.329376 4.007039 70.89744 1.879664 13.41487 40.31877 118.37584
#> 102 17.53896 -3.192460 4.033962 72.74848 1.851044 13.50500 39.22320 119.84238
#> 103 17.90601 -3.055314 4.118382 74.62756 1.879080 13.78763 37.55666 119.69029
#> 104 17.90168 -2.932237 4.117387 76.45093 1.823372 13.78430 36.97282 119.13067
#> 105 18.28541 -2.805199 4.205643 78.30687 1.855938 14.07976 36.79216 118.56774
#> 106 18.56736 -2.685623 4.270492 80.17170 1.864830 14.29686 37.56453 118.72700
#> 107 18.64136 -2.575650 4.287513 81.99873 1.827029 14.35385 37.02630 118.43884
#> 108 18.49822 -2.475103 4.254590 83.74057 1.741839 14.24363 36.79208 118.04159
#> 109 18.01104 -2.386522 4.142540 85.32288 1.582310 13.86850 35.44097 116.71210
#> 110 17.70145 -2.297428 4.071333 86.78996 1.467081 13.63011 32.42475 113.40751
#> 111 17.68062 -2.205285 4.066542 88.20965 1.419691 13.61408 29.51097 108.67890
#> 112 17.72160 -2.115436 4.075968 89.59713 1.387483 13.64563 30.35668 105.72625
#> 113 17.78510 -2.028691 4.090572 90.95837 1.361241 13.69452 29.48201 102.81585
#> 114 17.92253 -1.943744 4.122182 92.31065 1.352274 13.80035 29.61558 100.70427
#> 115 18.08467 -1.861652 4.159474 93.65973 1.349085 13.92520 31.66368 100.65600
#> 116 18.25537 -1.782674 4.198734 95.00763 1.347900 14.05663 40.19262 107.10493
#>            K   Dumps_H  Dumps_I         R       AG    Dumps_K       AI
#> 91  29.92013 0.4540962 55.27622 16.124266 7.094677 -4.1900924 9.029589
#> 92  30.48022 0.4993261 51.45261 15.202264 6.688996 -3.8236110 8.513268
#> 93  30.93278 0.4609543 47.88995 14.975233 6.589102 -3.5626561 8.386130
#> 94  31.64046 0.4758296 44.61715 15.222440 6.697874 -3.2728067 8.524566
#> 95  32.43420 0.4920911 41.61300 14.421502 6.345461 -3.0041435 8.076041
#> 96  33.39494 0.5388566 38.88484 13.126576 5.775693 -2.7281636 7.350882
#> 97  34.03439 0.5433521 36.37227 12.503840 5.501689 -2.5125676 7.002150
#> 98  34.62157 0.5454871 34.05641 13.234386 5.823130 -2.3158591 7.411256
#> 99  35.48354 0.5572382 31.93089 13.834935 6.087372 -2.1255259 7.747564
#> 100 36.34186 0.5811756 29.99218 14.183722 6.240838 -1.9387029 7.942884
#> 101 37.29463 0.6011706 28.22219 13.959389 6.142131 -1.7699966 7.817258
#> 102 37.75666 0.6086564 26.59628 13.015000 5.726600 -1.6259057 7.288400
#> 103 37.70875 0.6117572 25.09925 13.175455 5.797200 -1.4970279 7.378255
#> 104 37.53244 0.6028408 23.71003 12.128866 5.336701 -1.3892246 6.792165
#> 105 37.35509 0.5989865 22.42492 11.791060 5.188066 -1.2851070 6.602994
#> 106 37.40526 0.6006701 21.24097 11.519021 5.068369 -1.1839504 6.450652
#> 107 37.31447 0.5992122 20.14744 12.108304 5.327654 -1.0935321 6.780650
#> 108 37.18932 0.5972024 19.13681 11.913900 5.242116 -1.0106309 6.671784
#> 109 36.77046 0.5904762 18.19830 11.627654 5.116168 -0.9385059 6.511486
#> 110 35.72934 0.5737574 17.31711 11.150673 4.906296 -0.8811890 6.244377
#> 111 34.23958 0.5498342 16.48215 10.544143 4.639423 -0.8349607 5.904720
#> 112 33.30934 0.5348960 15.69813 10.257673 4.513376 -0.7840267 5.744297
#> 113 32.39241 0.5201715 14.96128  9.975303 4.389133 -0.7368432 5.586170
#> 114 31.72715 0.5094885 14.27169  9.770436 4.298992 -0.6895884 5.471444
#> 115 31.71194 0.5092443 13.63533  9.765753 4.296931 -0.6363665 5.468822
#> 116 33.74370 0.5418711 13.07838 10.391435 4.572231 -0.5569463 5.819203
#>            AJ          AL Total_swpstock_change_LD Total_paper_stock_change_LD
#> 91   89.21922  4.76462010                 10.52477                    7.669205
#> 92   93.27370  4.05448178                 11.67483                    6.919867
#> 93   97.02187  3.74816940                 10.14263                    6.774616
#> 94  100.73116  3.70928935                 10.37043                    7.134356
#> 95  103.84316  3.11200048                 10.58904                    6.453318
#> 96  106.12110  2.27794830                 11.64977                    5.325478
#> 97  107.96231  1.84120069                 11.52553                    4.830323
#> 98  110.10995  2.14764234                 11.35296                    5.654913
#> 99  112.48057  2.37062635                 11.38124                    6.332472
#> 100 114.92946  2.44888245                 11.80691                    6.751017
#> 101 117.14672  2.21726365                 11.96516                    6.589398
#> 102 118.75810  1.61137642                 12.16359                    5.712071
#> 103 120.38171  1.62361695                 12.61139                    5.923789
#> 104 121.37191  0.99019275                 12.67543                    4.937669
#> 105 122.13638  0.76447707                 13.13050                    4.667436
#> 106 122.72059  0.58420806                 13.47607                    4.468627
#> 107 123.59309  0.87249829                 13.60523                    5.106620
#> 108 124.32188  0.72879335                 13.51036                    4.960278
#> 109 124.86444  0.54255956                 13.06429                    4.720221
#> 110 125.12733  0.26288355                 12.79977                    4.287991
#> 111 125.05406 -0.07327070                 12.82848                    3.731192
#> 112 124.83102 -0.22303202                 12.91768                    3.506318
#> 113 124.46725 -0.36376998                 13.02707                    3.288520
#> 114 124.01059 -0.45666556                 13.20888                    3.152738
#> 115 123.57225 -0.43833412                 13.41263                    3.222231
#> 116 123.48831 -0.08393984                 13.62186                    3.931345
#>     Var1b_STOCKCHANGE_TOTAL
#> 91                 18.19397
#> 92                 18.59470
#> 93                 16.91725
#> 94                 17.50479
#> 95                 17.04236
#> 96                 16.97525
#> 97                 16.35585
#> 98                 17.00787
#> 99                 17.71371
#> 100                18.55793
#> 101                18.55456
#> 102                17.87566
#> 103                18.53518
#> 104                17.61310
#> 105                17.79794
#> 106                17.94470
#> 107                18.71185
#> 108                18.47064
#> 109                17.78451
#> 110                17.08776
#> 111                16.55967
#> 112                16.42400
#> 113                16.31559
#> 114                16.36162
#> 115                16.63486
#> 116                17.55320

calculatedumpcarbonstockchange(Years = 1990:2015, totalcarbonstockchange = TRUE)

#>  [1] 18.19397 18.59470 16.91725 17.50479 17.04236 16.97525 16.35585 17.00787
#>  [9] 17.71371 18.55793 18.55456 17.87566 18.53518 17.61310 17.79794 17.94470
#> [17] 18.71185 18.47064 17.78451 17.08776 16.55967 16.42400 16.31559 16.36162
#> [25] 16.63486 17.55320

calculatedumpcarbonstockchange(Years = 1990:2015, totalcarbonstockchange = TRUE)

#> [1] 18.30763 18.60157 17.00552 17.62658 17.22093 17.05147 16.34753 17.09010 #> [9] 17.76937 18.66230 18.50829 17.61024 18.23545 17.32596 17.64350 17.76522 #> [17] 18.58679 18.30797 17.51054 16.64159 16.30085 16.17128 16.11204 16.18834 #> [25] 17.14643 17.96544 #>

>>>>>>> master