Values for final 7 variables.

Values for final 7 variables. All units are in Gg Carbon Var1A = Annual Change in stock of HWP in use form consumption Var1B = Annual Change in stock of HWP in SWDS from consumption Var2A = Annual Change in stock of HWP in use produced from domestic harvest Var2B = Annual Change in stock of HWP in SWDS produced from domestic harvest Var3 = Annual Imports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips Var4 = Annual Exports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips Var5 = Annual Domestic Harvest Var6 = Annual release of carbon to the atmosphere from HWP consumption (from fuelwood & products in use and products in SWDS) Var 7 = Annual release of carbon to the atmosphere from HWP (including fuelwood) where wood came from domestic harvest (from products in use and products in SWDS)

finalVariables(Years = 1990:2015, Variable = paste("Var", c("1A", "1B",
  "2A", "2B", "3", "4", "5", "6", "7"), sep = ""),
  decaydistribution = c("Exponential", "K=2"), halflives = halfLives)

Arguments

Years

years to calculate

Variable

variable to return

decaydistribution

decay distribution to use (if necessary)

halflives

half-lives for calculation (if necessary)

Value

a numeric vector of values for the selected variable for `Years`

Examples

finalVariables(Variable = "Var1A")

#>  [1] 17049.8828 13695.6138 15189.4365 16753.2544 17683.7151 17822.5016
#>  [7] 17634.5446 18109.6498 19928.8670 21083.5096 21189.8535 18452.8143
#> [13] 17809.7296 17095.1977 20251.1521 21327.5882 18601.3882 13520.3504
#> [19]  3592.0678 -3267.9224 -4592.0846 -1859.2830  -708.6011  2062.6613
#> [25]  6099.9515 14343.3463

finalVariables(Variable = "Var2B")

#>  [1] 18131.23 18726.34 17260.19 17782.21 17248.91 17119.64 16452.02 17091.22
#>  [9] 17661.86 18255.52 18019.40 17022.70 17425.81 16395.06 16371.50 16413.09
#> [17] 17352.37 17489.53 17304.84 17069.91 16778.96 16998.70 17090.29 17158.66
#> [25] 17069.76 17496.19