Calculates the 7 HWP Variables.

Returns variables needed to calculate overall contribution.

finalVariables(Years = 1990:2015, decaydistribution = c("Exponential",
  "K=2", "K=10"), halflives = halfLives, paperHL = 2.53087281800454,
  approach = NULL, fsp = fracstrpanels, fnsp = fracnonstrpanels,
  fsawn = fracsawnwood, woodToCarbon = 4.535925e-07,
  paperToCarbon = 3.9008955e-07, swpSwdsNondegradable = 0.77,
  paperSwdsNondegradable = 0.44, swpLandfillDecay = 0.0300063714528115,
  paperLandfillDecay = 0.0478032538317204,
  swpDumpDecay = 0.0420089200339361, paperDumpDecay = 0.0840178400678722,
  swpdata)

Arguments

Years

years to calculate

decaydistribution

decay distribution to use. Defaults to exponential distribution.

halflives

data frame of half lives to use. Must have a column associated with each one of 13 end uses and half life values for 1900 until the maximum year of interest.

paperHL

half life value for paper

approach

approach to calculate corresponding final variables for. If null, all final variables are calculated and returned.

fsp

Fraction of structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fnsp

Fraction of non-structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fsawn

Fraction of sawnwood products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

woodToCarbon

Conversion factor to convert oven-dry tons of solid wood to Tg of Carbon.

paperToCarbon

Conversion factor to convert oven-dry tons of paper to Tg of Carbon.

swpSwdsNondegradable

Percent of solid-wood products in landfills and dumps that are not subject to decay.

paperSwdsNondegradable

Percent of paper products in landfills and dumps that are not subject to decay.

swpLandfillDecay

Decay rate of solid-wood products in landfills as a half-life in years.

paperLandfillDecay

Decay rate of paper products in landfills as a half-life in years.

swpDumpDecay

Decay rate of solid-wood products in dumps as a half-life in years.

paperDumpDecay

Decay rate of paper products in dumps as a half-life in years.

swpdata

Optional argument to provide swp production and trade time series for purposes of error analysis

Value

If an approach is specified then only the variables needed for that approach will be returned. If approach is null then a data frame with nine columns (Variables 1 and 2 have two parts each) will be returned. If the decay distribution is exponential then the data frame returned will correspond to table containing final variables: `06 IPCC Tables` in the WOODCARB II spreadsheet.

Details

All units are in Gg Carbon.

Var1A = Annual Change in stock of HWP in use form consumption

Var1B = Annual Change in stock of HWP in SWDS from consumption

Var2A = Annual Change in stock of HWP in use produced from domestic harvest

Var2B = Annual Change in stock of HWP in SWDS produced from domestic harvest

Var3 = Annual Imports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips

Var4 = Annual Exports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips

Var5 = Annual Domestic Harvest

Var6 = Annual release of carbon to the atmosphere from HWP consumption (from fuelwood & products in use and products in SWDS)

Var7 = Annual release of carbon to the atmosphere from HWP (including fuelwood) where wood came from domestic harvest (from products in use and products in SWDS)

Examples

finalVariables(Years = 1950:2000)

#>    Years     Var1A     Var1B     Var2A     Var2B      Var3      Var4      Var5
#> 1   1950  9995.945  6073.986  8325.377  5524.132        NA        NA  27403.16
#> 2   1951  8212.001  6923.282  7224.940  6369.621        NA        NA  26690.07
#> 3   1952  7216.375  7084.739  6456.358  6535.815        NA        NA  24420.43
#> 4   1953  7571.742  7320.026  6505.250  6742.153        NA        NA  24225.81
#> 5   1954  7114.300  7500.246  6151.799  6930.567        NA        NA  23937.36
#> 6   1955  8501.298  7838.078  7424.610  7273.041        NA        NA  24426.36
#> 7   1956  8656.789  8145.093  7440.233  7560.769        NA        NA  24053.28
#> 8   1957  5410.116  8175.901  4619.656  7620.178        NA        NA  22334.69
#> 9   1958  5599.659  8332.094  4669.402  7785.624        NA        NA  21731.01
#> 10  1959  7898.377  8657.372  6530.785  8081.612        NA        NA  22687.63
#> 11  1960  5669.688 10469.543  4665.256  9784.615        NA        NA  21886.08
#> 12  1961  5611.774 10530.369  4592.384  9872.030        NA        NA  21759.52
#> 13  1962  6783.654 10735.669  5294.830 10046.057        NA        NA  22407.38
#> 14  1963  8125.893 10986.572  6881.941 10323.849        NA        NA  22549.87
#> 15  1964  9660.151 11337.620  8455.472 10684.466        NA        NA  22370.50
#> 16  1965 11471.816 11820.256 10062.247 11124.109  6051.165  2809.271  94648.95
#> 17  1966 12055.957 12295.465 10688.924 11552.053  6443.170  3282.320  96992.34
#> 18  1967 10276.461 12533.487  9677.177 11849.818  6201.126  4212.135  95590.20
#> 19  1968 11746.328 13032.489 11137.073 12417.928  6893.801  5371.828  99324.39
#> 20  1969 12234.910 13564.390 11339.477 12989.341  7375.913  5686.274 100621.23
#> 21  1970 10866.400 13860.862 10596.254 13441.040  7115.558  6662.083 109529.37
#> 22  1971 12863.347 14391.855 11478.406 13952.717  7773.550  5679.069 103970.74
#> 23  1972 14901.016 15103.312 13168.105 14648.918  8943.554  6908.594 106269.04
#> 24  1973 15091.262 15811.116 13824.360 15397.897  9222.992  7893.552 107649.81
#> 25  1974 10210.188 15983.698  9941.003 15705.881  8250.261  7916.759 105143.84
#> 26  1975  4984.634 15576.003  5485.384 15495.069  6468.645  7152.180  96291.62
#> 27  1976 10545.801 16232.586 10212.110 16190.109  8136.783  8333.654 102263.65
#> 28  1977 13357.679 16892.739 11791.598 16802.009  9375.688  8173.078 107011.65
#> 29  1978 14886.408 17641.237 12303.038 17394.988 10578.096  8373.062 113275.68
#> 30  1979 14180.854 18331.608 12707.239 18114.098 10478.606  9828.946 118136.49
#> 31  1980  9142.947 18481.331  9239.598 18508.293  9469.013 10351.657 112410.20
#> 32  1981  7549.515 18121.049  6721.279 18164.861  9282.627  8872.322 108744.69
#> 33  1982  6376.251 17647.887  5238.875 17677.024  9141.002  8846.670 113021.70
#> 34  1983 13308.289 18072.427 10793.581 17930.769 11208.629  9564.347 124168.21
#> 35  1984 16657.593 18532.109 12948.507 18139.962 12779.331  9535.929 129787.76
#> 36  1985 16164.061 18642.347 11992.741 18026.413 13497.558  9823.927 128841.85
#> 37  1986 18637.505 18988.068 15337.012 18328.227 13588.834 10691.868 139143.20
#> 38  1987 21004.761 19436.830 17958.100 18739.661 14575.982 12102.269 140517.41
#> 39  1988 19885.065 19615.605 19096.707 19130.557 14099.278 14579.840 142284.64
#> 40  1989 18591.500 19650.691 18633.365 19363.136 13662.151 15253.874 142955.07
#> 41  1990 17043.712 18307.634 17659.334 18278.363 12679.503 15077.999 142297.14
#> 42  1991 13128.825 18601.567 14939.855 18812.331 11551.747 15667.472 144434.61
#> 43  1992 15717.739 17005.519 16334.429 17426.645 12855.636 16032.210 139388.54
#> 44  1993 16956.723 17626.583 14971.221 17949.011 14511.739 14788.173 134553.68
#> 45  1994 18220.973 17220.928 15929.570 17479.060 15684.969 15665.370 134750.37
#> 46  1995 17306.863 17051.466 15064.955 17228.990 16712.320 17265.886 137026.51
#> 47  1996 17018.001 16347.532 14092.141 16513.068 16690.669 16732.919 134477.27
#> 48  1997 18755.629 17090.104 14740.118 17262.744 17983.122 16876.897 135438.73
#> 49  1998 19653.677 17769.368 13403.948 17738.355 18994.153 15056.852 134206.11
#> 50  1999 21444.228 18662.301 14145.556 18358.728 20599.100 15245.335 134192.96
#> 51  2000 19999.515 18508.292 12840.060 17969.614 21857.503 16184.631 133694.05

finalVariables(halflives = halfLives * 1.25)

#>    Years      Var1A    Var1B       Var2A    Var2B     Var3     Var4      Var5
#> 1   1990 17043.7116 18307.63 17659.33357 18278.36 12679.50 15078.00 142297.14
#> 2   1991 13128.8249 18601.57 14939.85484 18812.33 11551.75 15667.47 144434.61
#> 3   1992 15717.7392 17005.52 16334.42946 17426.64 12855.64 16032.21 139388.54
#> 4   1993 16956.7233 17626.58 14971.22105 17949.01 14511.74 14788.17 134553.68
#> 5   1994 18220.9725 17220.93 15929.56956 17479.06 15684.97 15665.37 134750.37
#> 6   1995 17306.8633 17051.47 15064.95468 17228.99 16712.32 17265.89 137026.51
#> 7   1996 17018.0011 16347.53 14092.14081 16513.07 16690.67 16732.92 134477.27
#> 8   1997 18755.6290 17090.10 14740.11768 17262.74 17983.12 16876.90 135438.73
#> 9   1998 19653.6773 17769.37 13403.94806 17738.36 18994.15 15056.85 134206.11
#> 10  1999 21444.2277 18662.30 14145.55577 18358.73 20599.10 15245.33 134192.96
#> 11  2000 19999.5154 18508.29 12840.05979 17969.61 21857.50 16184.63 133694.05
#> 12  2001 16491.4294 17610.24  8713.21957 16780.98 22051.45 15336.17 127896.01
#> 13  2002 17413.8298 18235.45  9565.67340 17212.91 23209.92 15743.99 126866.06
#> 14  2003 16986.4185 17325.96  9452.50336 16174.97 23706.93 16302.59 126477.26
#> 15  2004 21409.1983 17643.50 12079.90850 16274.85 26427.58 16952.97 131737.71
#> 16  2005 20989.8919 17765.22 11710.88488 16293.62 26792.51 17312.49 132481.61
#> 17  2006 19082.7521 18586.79 12095.26266 17268.23 25442.39 18835.68 129528.83
#> 18  2007 13091.5426 18307.97 10639.00367 17387.19 21649.52 20656.59 123639.75
#> 19  2008  2420.1492 17510.54  3864.02978 17089.68 16981.71 21158.54 106096.03
#> 20  2009 -5103.8430 16641.59 -1820.66173 16646.38 13114.89 20616.42  96032.07
#> 21  2010 -2895.9387 16300.85   -59.18892 16572.71 14161.49 22419.91  97555.35
#> 22  2011 -1887.2390 16171.28  1998.34356 16804.35 13922.97 24671.92 100848.20
#> 23  2012  -298.5597 16112.04  1724.39575 16906.41 14067.44 23323.55 103469.51
#> 24  2013  2298.3609 16188.34  2318.94405 16994.18 15142.07 22850.52 107005.04
#> 25  2014 12943.6684 17146.43  6234.96310 17366.38 19008.74 18657.41  98861.15
#> 26  2015 13474.0869 17965.44  7141.21700 17745.38 19872.28 18716.93 100692.34