Timber products: Production, imports, exports, and new supply of lumber in the United States, specified years, 1799-1958

Dataset containing statistics for timber products. All units are in Million Board Ft, unless otherwise specified. All vectors are numeric.

hair1958

Format

A data frame with 58 observations on 23 variables

Prod.Tot Production Total

Prod.SW Production of Softwoods

Prod.HW Production of Hardwoods

Imports.Tot Total Imports

Imports.SW Imports for Softwood

Imports.HW Imports for Hardwoods

Imports.Mixed Imports for Mixed Woods

Imports.Mixed.PercOfTot Mixed Woods Percent of Total Imports

Imports.SW.PerOfTot Softwoods Percent of Total Imports

Imports.HW.PercOfTot Hardwoods Percent of Total Imports

Imports.Estimated.SW Estimated Softwood Imports

Imports.Estimated.HW Estimated Hardwood Imports

Exports.Tot Total Exports

Exports.SW Softwood Exports

Exports.HW Hardwood Exports

Exports.Mixed Mixed Woods Exports

Exports.Mixed.PercOfTot Mixed Woods Percent of Total Exports

Exports.SW.PercOfTot Softwoods Percent of Total Exports

Exports.HW.PercOfTot Hardwoods Percent of Total Exports

Exports.Estimated.SW Estimated Softwood Exports

Exports.Estimated HW Estimated Hardwood Exports

NewSupply New Supply

PerCapita Per Capita Use, in Board ft.

Source

USFS Estimations and Bureau of Census Data