Total carbon in solid-wood products (od tons of wood fiber)

Calculates carbon per end use and total carbon in solidwood products.

swpcarbontotal(Yrs = 1990:2015, distribution = c("Exponential", "K=2"),
  onlytotal = TRUE, lumberpre = TRUE, approach = c("Production",
  "Stock Change"), halflives = halfLives)

Arguments

Yrs

years to calculate carbon totals for

distribution

type of decay distribution

onlytotal

return total carbon or totals by end uses

lumberpre

whether to include carbon from lumber before 1900 in totals

approach

approach type to use to calculate

halflives

data frame of half lives for end uses. Default half lives are used but a data frame with 13 columns with half lives for the appropriate years can be used.

Value

returns either only total for all end uses or total by enduses with option of lumberpre1900

Details

ALl units returned are in od tons of wood fiber. If "onlytotal" is false, total carbon is returned for each one of the 13 end uses. If "onlytotal" is true, the totals are summed up and one total is returned. test

This corresponds to the `SWCalcP$H4389` column in the WOODCARB spreadsheet for Production approach. This corresponds to the `SWCalc$H4389` column in the WOODCARB spreadsheet for Stock Change approach.

Examples

swpcarbontotal(onlytotal=FALSE)

#>    Years       EU1       EU2      EU3       EU4       EU5      EU6     EU7
#> 1   1990 706414309 182866038 63065693 258801895 116211294 32956528 3874448
#> 2   1991 718762376 182025862 64273246 265839670 118331002 32929783 3892945
#> 3   1992 731678400 181261960 65527205 272864493 120584443 32900267 3913323
#> 4   1993 744388871 180504630 66736436 279331347 122805240 32846192 3931937
#> 5   1994 757463259 179806339 67970042 285680155 125121464 32787124 3951882
#> 6   1995 770109737 179085699 69111065 291258590 127317669 32699310 3969296
#> 7   1996 783085062 178419160 70269949 296655311 129591510 32600626 3987234
#> 8   1997 796265175 177793137 71423069 301843229 131916131 32498620 4006674
#> 9   1998 809303685 177174566 72529660 306536730 134202034 32377195 4024942
#> 10  1999 822842157 176623303 73690647 311536504 136597079 32276103 4046224
#> 11  2000 836196543 176074498 74824810 316329446 138929063 32171235 4066189
#> 12  2001 848372870 175429599 75795248 320278298 140954071 32045972 4081411
#> 13  2002 860932376 174841493 76811896 324425431 143057527 31930964 4098033
#> 14  2003 872990427 174224761 77761738 328180039 145025174 31806474 4112205
#> 15  2004 885911212 173708487 78810771 332516025 147193781 31712971 4131430
#> 16  2005 898871507 173213089 79855288 336826401 149354021 31624902 4151041
#> 17  2006 911264634 172682063 80824803 340723825 151367887 31528726 4168356
#> 18  2007 922375761 172046985 81631189 343694256 153056070 31401887 4179281
#> 19  2008 929859836 171091323 82007884 344150048 153869444 31177724 4171287
#> 20  2009 934859474 169924169 82091236 343100379 154123979 30912390 4154036
#> 21  2010 940141701 168801706 82205622 342217459 154428853 30651352 4137418
#> 22  2011 946195215 167771842 82412718 341994687 154937218 30427706 4126859
#> 23  2012 952610811 166793130 82658092 342018146 155520332 30215707 4118118
#> 24  2013 959991207 165920706 83009756 342693228 156318555 30032268 4114394
#> 25  2014 967726682 165102229 83411477 343661119 157212483 29864745 4113017
#> 26  2015 976024693 164356403 83886558 345051002 158248655 29717939 4114937
#>          EU8      EU9      EU10      EU11      EU12      EU13 LumberPre1900
#> 1   97501711 32435774  83160265 180886699 231453156 478162361      70903425
#> 2   99435437 33950971  84995306 182322663 232000891 476933616      69188935
#> 3  101537077 35618738  86929782 183884013 232858760 475666020      67515904
#> 4  103699417 37413398  88857437 185382430 233859425 474028383      65883327
#> 5  106040918 39391552  90875835 186987316 235167761 470859419      64290227
#> 6  108288092 41311366  92770246 188477472 236466435 467559978      62735649
#> 7  110684023 43410498  94739118 190024786 238028850 463715374      61218661
#> 8  113202054 45594228  96742737 191693368 239853120 459779274      59738356
#> 9  115745943 47839549  98716186 193327188 241790971 455578388      58293845
#> 10 118402832 50136072 100773518 195123787 243917993 451359631      56884263
#> 11 120999083 52418301 102787989 196832870 245974176 447244425      55508766
#> 12 123235010 54365656 104521210 198262298 247598480 443467109      54166529
#> 13 125560790 56409340 106329705 199758848 249353058 439730218      52856749
#> 14 127744638 58360137 108029062 201105311 250934456 436068692      51578640
#> 15 130132157 60407160 109881902 202732116 252833676 432473766      50331436
#> 16 132494509 62411997 111718514 204369940 254729925 428944236      49114390
#> 17 134710044 64304201 113435197 205870110 256452637 425505854      47926773
#> 18 136582743 65970465 114887793 206995933 257724690 422197382      46767874
#> 19 137620435 67222012 115683937 207037612 257814750 418477723      45636997
#> 20 138155370 68103392 116040238 206564495 257255109 414872819      44533466
#> 21 138693102 68985296 116424536 206107483 256716678 411340193      43456619
#> 22 139433967 69905761 116950161 205994505 256517523 407871802      42405811
#> 23 140207663 70811834 117516753 205970074 256395519 404466488      41380412
#> 24 141137932 71748596 118223611 206215335 256535014 401123110      40379808
#> 25 142191046 72778943 119017913 206589579 256857165 397921419      39403399
#> 26 143407013 73922231 119933477 207145928 257419209 394777995      38450600
#>    Total Carbon
#> 1    2538693597
#> 2    2564882702
#> 3    2592740383
#> 4    2619668468
#> 5    2646393294
#> 6    2671160603
#> 7    2696430161
#> 8    2722349173
#> 9    2747440884
#> 10   2774210114
#> 11   2800357395
#> 12   2822573762
#> 13   2846096429
#> 14   2867921752
#> 15   2892776889
#> 16   2917679760
#> 17   2940765110
#> 18   2959512312
#> 19   2965821012
#> 20   2964690550
#> 21   2964308018
#> 22   2966945774
#> 23   2970683079
#> 24   2977443519
#> 25   2985851216
#> 26   2996456638

swpcarbontotal(1950:1975, approach = "Stock Change")

#>  [1] 1818871465 1833139774 1846854799 1860447189 1873679776 1888553014
#>  [7] 1903747351 1913075329 1923424778 1937991525 1948188912 1957986752
#> [13] 1969760420 1984637567 2001804356 2022063587 2042473059 2060743598
#> [19] 2081491502 2102594227 2122853013 2147675904 2174928424 2201899977
#> [25] 2220103286 2233672654